% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_citation.R
\name{add_citation}
\alias{add_citation}
\title{Add a citation to a design}
\usage{
add_citation(design, title = NULL, author = NULL, year = NULL,
  description = "Unpublished research design declaration", citation = NULL)
}
\arguments{
\item{title}{(optional) The title of the design, as a character string.}

\item{year}{(optional) The year of the design, as a character string.}

\item{description}{(optional) A description of the design in words, as a character string.}

\item{citation}{(optional) The preferred citation for the design, as a character string.}

\item{authors}{(optional) The authors of the design, as a character string.}
}
\value{
a design object with a citation attribute
}
\description{
Add a citation to a design
}
\examples{

design <- 
declare_population(data = sleep) + 
  declare_sampling(n = 10)

design <- 
  add_citation(design,
               author = "Lovelace, Ada",
               title = "Notes",
               year = 1953,
               description = "This is a text description of a design")

cite_design(design)

}
