% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reveal_design.R
\name{assign_treatment}
\alias{assign_treatment}
\title{Assign treatment status}
\usage{
assign_treatment(design, data)
}
\arguments{
\item{design}{A design object created by \code{\link{declare_assignment}}; or a function that assigns treatment}

\item{data}{A dataframe, often created by \code{\link{draw_population}} or \code{\link{draw_sample}}.}
}
\value{
A random assignment vector of length N.
}
\description{
Description
}
\examples{
# these examples don't work yet
# smp <- declare_population(N = 850)
# po <- declare_potential_outcomes(condition_names = c("Z0","Z1"),
#                                    outcome_formula = Y ~ .01 + 0*Z0 + .2*Z1)
# design <- declare_assignment(potential_outcomes = po, m=200)
# mock          <- draw_population(potential_outcomes = po, sample =  smp)
# Z        <- assign_treatment(design, data = mock)
# table(Z)
}

