% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_treatment.R
\name{assign_treatment_indicator}
\alias{assign_treatment_indicator}
\title{Assign treatment status}
\usage{
assign_treatment_indicator(data, assignment, random_seed = NULL)
}
\arguments{
\item{data}{A dataframe, often created by \code{\link{draw_population}} or \code{\link{draw_sample}}.}

\item{assignment}{A assignment object created by \code{\link{declare_assignment}}; or a function that assigns treatment}
}
\value{
A random assignment vector of length N.
}
\description{
This function takes a data.frame and an assignment object and returns an assignment vector.  Users will often prefer to use \code{\link{assign_treatment}}.
}
\details{
Description
}
\examples{
population <- declare_population(noise = declare_variable(), N = 1000)
sampling <- declare_sampling(n = 500)
potential_outcomes <- declare_potential_outcomes(formula = Y ~ 5 + .5*Z + noise,
                                                 condition_names = c(0, 1),
                                                 treatment_variable = "Z")
assignment <- declare_assignment(condition_names = c(0,1))
pop_draw <- draw_population(population = population)
smp_draw <- draw_sample(data = pop_draw, sampling = sampling)
Z <- assign_treatment_indicator(data = smp_draw, assignment=assignment)
table(Z)
}

