% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_functions.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Bootstrap data, including multi-level data}
\usage{
bootstrap_data(data, size, level_IDs = NULL)
}
\arguments{
\item{data}{A data.frame, usually provided by the user and not created with the DeclareDesign package.}

\item{N}{The number of units in the resulting bootstrapped data.frame.}

\item{N_per_level}{The number of units in each level of the bootstrapped heirarchy.}

\item{group_sizes_per_level}{A list of group sizes per level in the heirarchy.}

\item{level_ID_variables}{A vector of variable names, indicating the variables that desibe the heirarchy.}
}
\value{
A data.frame
}
\description{
Bootstrap data, including multi-level data
}

