% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_functions.R
\name{calculate_ICC}
\alias{calculate_ICC}
\title{Estimate the ICC of a given variable}
\usage{
calculate_ICC(variable, cluster_variable)
}
\arguments{
\item{variable}{A vector whose intra-cluster correlation coefficient should be estimated}

\item{cluster_variable}{A vector of cluster IDs indicating the cluster corresponding to each element of variable}
}
\description{
Estimate the ICC of a given variable
}

