% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_design.R
\name{check_design}
\alias{check_design}
\title{Check design for consistency between design and analysis}
\usage{
check_design(design, covariates, potential_outcomes, analysis)
}
\arguments{
\item{design}{design object}

\item{covariates}{covariates object}

\item{potential_outcomes}{potential outcomes object}

\item{analysis}{analysis object}

\item{sims}{number of iterations}
}
\description{
Check design for consistency between design and analysis
}

