% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_registration.R
\name{check_registration}
\alias{check_registration}
\title{Check a pre-registration for consistency between design and analysis}
\usage{
check_registration(design, covariates, potential_outcomes, analysis)
}
\arguments{
\item{design}{design object}

\item{covariates}{covariates object}

\item{potential_outcomes}{potential outcomes object}

\item{analysis}{analysis object}

\item{sims}{number of iterations}
}
\description{
Check a pre-registration for consistency between design and analysis
}

