% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_design_code}
\alias{compare_design_code}
\title{Compare the code of two designs}
\usage{
compare_design_code(...)
}
\arguments{
\item{...}{Options sent to \code{diffobj::diffObj}}

\item{design1}{A design object, typically created using the + operator}

\item{design2}{A design object, typically created using the + operator}
}
\value{
A Diff object that can be compared using the diffobj package
}
\description{
Compare the code of two designs
}
