% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_design_from_template.R
\name{compare_designs_from_template}
\alias{compare_designs_from_template}
\title{Compare Designs from Template Function}
\usage{
compare_designs_from_template(template, ...)
}
\arguments{
\item{template}{Template function to create a set of designs to compare}

\item{...}{options for the template function over which to compare designs}
}
\description{
Compare Designs from Template Function
}

