% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare_experiments.R
\name{compare_experiments}
\alias{compare_experiments}
\title{Declare the data-generating process of a variable}
\usage{
compare_experiments(N = NULL, N_per_level = NULL,
  group_sizes_per_level = NULL, design = NULL, analysis = NULL,
  sample = NULL, potential_outcomes = NULL, blocks = NULL,
  clusters = NULL, sims = 5, labels = NULL, analysis_labels = NULL)
}
\arguments{
\item{N}{what is it?}

\item{N_per_level}{what is it?}

\item{group_sizes_per_level}{what is it?}

\item{design}{what is it?}

\item{analysis}{what is it?}

\item{sample}{what is it?}

\item{potential_outcomes}{what is it?}

\item{blocks}{what is it?}

\item{clusters}{what is it?}

\item{sims}{what is it?}

\item{labels}{labels for each simulation}

\item{analysis_labels}{labels for each analysis}
}
\description{
Declare the data-generating process of a variable
}
\examples{
## here are examples
}

