% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_inputs.R
\name{compare_inputs}
\alias{compare_inputs}
\title{Compare the data and code inputs for two designs}
\usage{
compare_inputs(inputs)
}
\arguments{
\item{inputs}{A list of two input objects}
}
\description{
Compare the data and code inputs for two designs
}

