% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_sample.R
\name{covariates_table}
\alias{covariates_table}
\title{Print a table describing covariates described in sample}
\usage{
covariates_table(sample)
}
\arguments{
\item{sample}{a sample object created with declare_sample frame.}
}
\description{
Print a table describing covariates described in sample
}

