% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{currydata}
\alias{currydata}
\title{Curry data arguments into a handler function}
\usage{
currydata(FUN, dots)
}
\arguments{
\item{FUN}{A function that will be called with \code{data} and the arguments in \code{dots}.}

\item{dots}{A list of quosures/arguments to be spliced into \code{FUN} when invoked.}
}
\description{
Returns a function of one argument \code{data} that, when called,
evaluates \code{FUN} with the captured \code{dots} and with \code{data} supplied as needed.
}
\examples{
# DeclareDesign:::currydata(fabricatr::fabricate, list(N = 3, X = 1))
}
\keyword{internal}
