% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_analysis.R
\name{declare_analysis}
\alias{declare_analysis}
\title{Declare an experimental analysis}
\usage{
declare_analysis(formula, treatment_variable = "Z", method = "lm",
  subset = NULL, formula_estimand = NULL, weights_variable = NULL,
  qoi = "ATE", ...)
}
\arguments{
\item{formula}{an optional formula object to define analyses such as linear regressions with covariates}

\item{method}{either string (i.e. "lm" or "glm") or a function object that takes as arguments data, design and spits out a standard R class such as lm or glm}

\item{...}{additional options to be sent to the analysis function and the test_success function}

\item{test_success}{a function that extracts the binary result of a statistical test (did it pass = 1, if not = 0)}
}
\value{
a list containing a function to conduct the analysis and a function to extract the result of the test
}
\description{
Description
}
\examples{
# these examples don't work yet
# declare_analysis(analysis = "diff-in-means")
# declare_analysis(analysis = function(Y, Z, data) lm(paste(Y, "~", Z), data = data))
}

