% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_estimand.R
\name{declare_ATE}
\alias{declare_ATE}
\title{Built-in Estimand Functions: declare_ATE()}
\usage{
declare_ATE(condition_treat = "Z1", condition_control = "Z0",
  outcome = "Y", sep = "_")
}
\arguments{
\item{condition_treat}{The condition name of the "treatment" condition. By default "Z1".}

\item{condition_control}{The condition name of the "control" condition. By default "Z0".}

\item{outcome}{The name of the outcome variable.  By default "Y".}

\item{sep}{The separator used to create potential outcomes columns. By default "_", so that potential outcomes are constructed as, for example,  "Y_Z1" and "Y_Z0".}
}
\description{
This function is designed to be passed to the estimand argument of the \code{\link{declare_estimand}} function, and can be used to specify a common estimand, the average difference between two potential outcomes.
}

