% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_blocks.R
\name{declare_blocks}
\alias{declare_blocks}
\title{Declare the data-generating process of a variable}
\usage{
declare_blocks(blocks = NULL, recode = TRUE, recode_function = NULL,
  block_count = 5, clusters = NULL, blocks_function = NULL,
  block_name = "block_variable")
}
\arguments{
\item{blocks}{A character vector describing the variables in the sample dataframe that will be used for blocking.}

\item{recode}{logical.  If TRUE (the default), principal components analysis used to make equally sized blocks, where the number of blocks is defined by block_count}

\item{recode_function}{A user-supplied function that creates blocks. If NULL (the default), principal components will be used to make blocks.}

\item{block_count}{The total number of blocks.  If the blocking variables imply fewer than this number, nothing changes. If the variables supplied in blocks imply more possible categories, then the variable is recoded (if recode=TRUE).}

\item{clusters}{desc}

\item{blocks_function}{Function to take multiple variables and construct blocks from them. Must take blocks, sample, and block_name and return}

\item{block_name}{A character string that gives the name of the blocking variable.}
}
\description{
Declare the data-generating process of a variable
}

