% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_clusters.R
\name{declare_clusters}
\alias{declare_clusters}
\title{Declare cluster variable}
\usage{
declare_clusters(clusters, cluster_name = "cluster_variable",
  custom_cluster_function = NULL)
}
\arguments{
\item{clusters}{A character vector describing the variable in the covariates dataframe that will be used for clustering}

\item{cluster_name}{A character string that gives the name of the cluster variable.}

\item{custom_cluster_function}{A function that takes only a dataframe of covariates and returns a vector of length n whose entries describe which cluster each unit belongs to.}
}
\description{
Description
}

