% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_covariates.R
\name{declare_covariates}
\alias{declare_covariates}
\title{Declare the structure of the covariates}
\usage{
declare_covariates(..., N_per_level = NULL, lower_units_per_level = NULL)
}
\arguments{
\item{...}{a list either of variable declarations, or of lists of variable declarations (one per level)}
}
\description{
Declare the structure of the covariates
}

