% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_data_strategy.R
\name{declare_data_strategy}
\alias{declare_data_strategy}
\alias{data_strategy_handler}
\title{Declare data strategy}
\usage{
declare_data_strategy(..., handler = data_strategy_handler, label = NULL)

data_strategy_handler(data, ...)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}

\item{data}{A data.frame.}
}
\value{
A function that returns a data.frame.
}
\description{
Declare data strategy
}
\examples{

}
