% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_design_from_template.R
\name{declare_design_from_template}
\alias{declare_design_from_template}
\title{Declare Design from Template Function}
\usage{
declare_design_from_template(template, ...)
}
\arguments{
\item{template}{Template function to create a design.}

\item{...}{options for the template function.}
}
\description{
Declare Design from Template Function
}

