% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_DGP.R
\name{declare_DGP}
\alias{declare_DGP}
\title{Declare the data-generating process of a variable}
\usage{
declare_DGP(linear_mean = NULL, linear_sd = NULL,
  binary_probability = NULL, binary_categories = NULL,
  multinomial_probabilities = NULL, multinomial_categories = NULL)
}
\arguments{
\item{linear_mean}{If the variable is linear}

\item{variable_name}{The name of your new variable}
}
\description{
Declare the data-generating process of a variable
}

