% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_inputs.R
\name{declare_inputs}
\alias{declare_inputs}
\title{Draw a population data frame based on a declared data structure}
\usage{
declare_inputs(data_list = NULL, data_file = NULL, code_file = NULL)
}
\arguments{
\item{data_list}{List of data objects required to replicate a design}

\item{data_file}{File that includes data objects, in .RData format}

\item{code_file}{File of code, including functions used in a design}
}
\value{
An inputs object, including data and code required to replicate a design
}
\description{
Draw a population data frame based on a declared data structure
}

