% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_interference.R
\name{declare_interference}
\alias{declare_interference}
\title{Declare interference}
\usage{
declare_interference(exposure_function = default_exposure_function,
  formula = NULL, exposure_variable_name = NULL, condition_names,
  sep = "_", description = NULL, ...)
}
\arguments{
\item{exposure_function}{Function to reveal exposure to treatment}

\item{formula}{Formula used in exposure function}

\item{exposure_variable_name}{Name of variable indicating realized exposure to treatment}

\item{condition_names}{Condition name vector}

\item{sep}{Separator for naming exposure potential outcomes}

\item{description}{Describe the interference pattern declared}

\item{...}{A set of additional options for the \code{exposure_function}.}
}
\description{
Declare interference
}

