% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{declare_internal_inherit_params}
\alias{declare_internal_inherit_params}
\title{Inherit roxygen docs}
\usage{
declare_internal_inherit_params(..., handler = function(data, ...)
  data.frame(BLNK = "MSG"), label = NULL)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}
}
\description{
Inherit roxygen docs
}
\keyword{internal}
