% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_design.R
\name{declare_named_diagnosands}
\alias{declare_named_diagnosands}
\title{Helper to declare a set of named diagnosands, x}
\usage{
declare_named_diagnosands(x, alpha = 0.05, subset = NULL)
}
\description{
Helper to declare a set of named diagnosands, x
}
