% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_outcomes.R
\name{declare_reveal}
\alias{declare_reveal}
\title{Declare a Reveal Outcomes step}
\usage{
declare_reveal(..., handler = reveal_outcomes, label = NULL)
}
\arguments{
\item{...}{arguments for the handler}

\item{handler}{a handler function}

\item{label}{a step label}
}
\description{
Declare a Reveal Outcomes step
}
