% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_sample_frame.R
\name{declare_sample_frame}
\alias{declare_sample_frame}
\title{Declare the structure of the sample frame}
\usage{
declare_sample_frame(..., N_per_level = NULL, lower_units_per_level = NULL,
  N = NULL)
}
\arguments{
\item{...}{a list either of variable declarations, or of lists of variable declarations (one per level)}

\item{N_per_level}{description}

\item{lower_units_per_level}{description}
}
\description{
Declare the structure of the sample frame
}

