% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_DGP_functions.R
\name{declare_variable}
\alias{declare_variable}
\title{Declare the data-generating process of a variable}
\usage{
declare_variable(type = NULL, location_scale = NULL, min_max = NULL,
  probabilities = NULL, outcome_categories = NULL)
}
\arguments{
\item{type}{declare the type of variable}

\item{location_scale}{a vector containing the first and / or second moments of the distribution, if the variable takes mean and variance type arguments}

\item{min_max}{for uniform distributions, the minimum and maximum values}

\item{probabilities}{for binary and categorical variables, a vector of probabilities}

\item{outcome_categories}{for binary and categorical variables, labels to attach to outcomes}
}
\description{
Declare the data-generating process of a variable
}

