% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{DeclareDesign-deprecated}
\alias{DeclareDesign-deprecated}
\alias{declare_design}
\title{Deprecated functions}
\usage{
declare_design(...)
}
\description{
Deprecated functions
}
\section{For \code{declare_design}}{

\code{declare_design}, instead use the \code{\link{`+.dd`}} operator to create a design, i.e. \code{pop + smp + assgn}.
}

