% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeclareDesign.R
\docType{package}
\name{DeclareDesign}
\alias{DeclareDesign}
\alias{DeclareDesign-package}
\title{DeclareDesign package}
\description{
\begin{enumerate}
\item \code{\link{declare_assignment}}
\end{enumerate}
}
