% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_functions.R
\name{default_exposure_function}
\alias{default_exposure_function}
\title{Default potential outcomes function}
\usage{
default_exposure_function(formula, data, ...)
}
\arguments{
\item{formula}{A formula describing the outcomes as a function of a}

\item{data}{A data frame}

\item{options}{A named list (optional)}
}
\description{
Default potential outcomes function
}

