% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{get_estimands}
\title{Deprecated functions}
\usage{
get_estimands(...)
}
\arguments{
\item{...}{options sent to the old version of \code{get_estimands}.}
}
\description{
The function \code{get_estimands} has been replaced with \code{draw_estimands}.
}
