% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Diagnose the properties of a research design}
\usage{
diagnose(design = NULL, simulations = NULL,
  statistics = list(calculate_PATE, calculate_sd_SATE, calculate_power,
  calculate_RMSE, calculate_bias, calculate_coverage, calculate_type_S_rate,
  calculate_exaggeration_ratio), labels = c("PATE", "sd(SATE)", "Power",
  "RMSE", "Bias", "Coverage", "Type S Rate", "Exagg. Ratio"), sims = 5, ...)
}
\arguments{
\item{design}{}

\item{simulations}{}

\item{statistics}{}

\item{labels}{}

\item{sims}{}

\item{...}{}
}
\description{
Diagnose the properties of a research design
}

