% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_design.R
\name{download_design}
\alias{download_design}
\title{Get Design from the DeclareDesign Library of Research Designs}
\usage{
download_design(name)
}
\arguments{
\item{name}{Name of the design}
}
\value{
A design object or template function
}
\description{
Get Design from the DeclareDesign Library of Research Designs
}

