% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_design.R
\name{download_template}
\alias{download_template}
\title{Get Design Template from the DeclareDesign Library of Research Designs}
\usage{
download_template(name)
}
\arguments{
\item{name}{Name of the design template}
}
\value{
An R function
}
\description{
Get Design Template from the DeclareDesign Library of Research Designs
}
