% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pre_register_paper_draft.R
\name{draft_paper_from_pre_register}
\alias{draft_paper_from_pre_register}
\title{Create paper draft from a pre_registration}
\usage{
draft_paper_from_pre_register(pre_registration, dir = getwd(),
  temp_dir = FALSE, format = "rmarkdown", make_output = TRUE,
  keep_tex = FALSE, save_r_code = FALSE, output_format = "pdf",
  open_output = TRUE)
}
\arguments{
\item{pre_registration}{Object created by pre_register() function. The inputs to create a paper draft are extracted exactly from the pre-registered descriptions of the experiment.}

\item{make_output}{Indicator for whether code for registration document is compiled into a PDF, Microsoft Word, or HTML document}

\item{output_format}{String indicating which type of output file is created, "pdf" (default), "word", or "html"}

\item{open_output}{Indicator for whether the output file is opened after it is compiled.}

\item{file}{File name where object is saved.}

\item{type}{Type of document that is created, either \code{knitr} or \code{rmarkdown}, the default.}

\item{check_registration}{Indicates whether the design is evaluated for consistency with declared analyses.}

\item{...}{Other options for the render() command to create the output file from the markdown code.}
}
\value{
Filename and location where .Rmd or .Rnw and PDF file are saved.
}
\description{
Create paper draft from a pre_registration
}

