% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_register_paper_draft.R
\name{draft_paper_from_pre_registration}
\alias{draft_paper_from_pre_registration}
\title{Create paper draft from a pre_registration}
\usage{
draft_paper_from_pre_registration(pre_registration, data, dir = getwd(),
  temp_dir = FALSE, format = "rmarkdown", make_output = TRUE,
  keep_tex = FALSE, save_r_code = FALSE, output_format = "pdf",
  open_output = TRUE)
}
\arguments{
\item{pre_registration}{Object created by pre_register() function. The inputs to create a paper draft are extracted exactly from the pre-registered descriptions of the experiment.}

\item{data}{what is it?}

\item{dir}{what is it?}

\item{temp_dir}{what is it?}

\item{format}{Type of document that is created, either \code{knitr} or \code{rmarkdown}, the default.}

\item{make_output}{Indicator for whether code for registration document is compiled into a PDF, Microsoft Word, or HTML document}

\item{keep_tex}{what is it?}

\item{save_r_code}{what is it?}

\item{output_format}{String indicating which type of output file is created, "pdf" (default), "word", or "html"}

\item{open_output}{Indicator for whether the output file is opened after it is compiled.}
}
\value{
what is it?
}
\description{
Create paper draft from a pre_registration
}

