% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_blocks_draw_clusters.R
\name{draw_blocks}
\alias{draw_blocks}
\title{Draw blocks for custom block functions used in an assignment function}
\usage{
draw_blocks(data, assignment)
}
\arguments{
\item{data}{A data.frame used to draw blocks from.}

\item{assignment}{An \code{assignment} object created by \code{\link{declare_assignment}}.}
}
\value{
a data.frame including blocks.
}
\description{
Draw blocks for custom block functions used in an assignment function
}

