% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_blocks_draw_clusters.R
\name{draw_clusters}
\alias{draw_clusters}
\title{Draw clusters for custom cluster functions used in an assignment function}
\usage{
draw_clusters(data, assignment, random_seed = NULL)
}
\arguments{
\item{data}{}

\item{assignment}{}

\item{random_seed}{}
}
\value{
thing
}
\description{
Draw clusters for custom cluster functions used in an assignment function
}

