% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_outcomes.R
\name{draw_observed_outcome}
\alias{draw_observed_outcome}
\title{Draw observed outcome}
\usage{
draw_observed_outcome(data, potential_outcomes, condition_names = NULL)
}
\arguments{
\item{data}{data.frame}

\item{potential_outcomes}{A potential_outcomes object created by \code{\link{declare_potential_outcomes}}.}

\item{condition_names}{A vector of condition names.}
}
\value{
data.frame including the observed outcome
}
\description{
Draw observed outcome
}

