% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_outcomes.R
\name{draw_outcome}
\alias{draw_outcome}
\title{Draw observed outcome}
\usage{
draw_outcome(data, potential_outcomes, condition_names = NULL,
  attrition = NULL)
}
\arguments{
\item{data}{}

\item{potential_outcomes}{}
}
\description{
Draw observed outcome
}

