% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_outcomes.R
\name{draw_outcome_vector}
\alias{draw_outcome_vector}
\title{Draw observed outcome (vector)}
\usage{
draw_outcome_vector(data, potential_outcomes, attrition = NULL)
}
\arguments{
\item{data}{}

\item{potential_outcomes}{}
}
\description{
Draw observed outcome (vector)
}

