% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_population.R
\name{draw_population}
\alias{draw_population}
\title{Make the full dataset or just a sample}
\usage{
draw_population(population, potential_outcomes = NULL,
  condition_names = NULL)
}
\description{
Make the full dataset or just a sample
}

