% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_outcomes.R
\name{draw_potential_outcome_vector}
\alias{draw_potential_outcome_vector}
\title{Draw potential outcome vector}
\usage{
draw_potential_outcome_vector(data, potential_outcomes, condition_name)
}
\arguments{
\item{data}{}

\item{potential_outcomes}{}

\item{condition_name}{}

\item{assignment_variable_name}{}
}
\description{
Draw potential outcome vector
}

