% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_outcomes.R
\name{draw_potential_outcomes}
\alias{draw_potential_outcomes}
\title{Draw potential outcomes}
\usage{
draw_potential_outcomes(data, condition_names = NULL, potential_outcomes,
  interference = NULL, noncompliance = NULL, attrition = NULL)
}
\arguments{
\item{data}{}

\item{condition_names}{}

\item{potential_outcomes}{}
}
\description{
Draw potential outcomes
}

