% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample.R
\name{draw_sample_indicator}
\alias{draw_sample_indicator}
\title{draw sample from population}
\usage{
draw_sample_indicator(data, sampling, random_seed = NULL)
}
\arguments{
\item{data}{A dataframe, often created by \code{\link{draw_population}}.}

\item{sampling}{A sampling object created by \code{\link{declare_sampling}}; or a function that samples}
}
\value{
A vector of 0's and 1's indicating which population units are sampled.
}
\description{
Description
}

