% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{error_in_draw}
\alias{error_in_draw}
\title{Run through draw_ functions to detect errors at any step}
\usage{
error_in_draw(design, message_index)
}
\description{
Run through draw_ functions to detect errors at any step
}

