% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ICC.R
\name{estimate_ICC}
\alias{estimate_ICC}
\title{Estimate the ICC of a given variable}
\usage{
estimate_ICC(variable, cluster)
}
\arguments{
\item{variable}{A vector whose intra-cluster correlation coefficient should be estimated}

\item{cluster}{A vector of cluster IDs indicating the cluster corresponding to each element of variable}
}
\description{
Estimate the ICC of a given variable
}

