% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{execution_st}
\alias{execution_st}
\title{Build an execution strategy object}
\usage{
execution_st(design, current_df = NULL, results = NULL, start = 1,
  end = length(design))
}
\arguments{
\item{design}{a design}

\item{current_df}{a data.frame}

\item{results}{a list of intermediate results}

\item{start}{index of starting step}

\item{end}{index of ending step}
}
\description{
Build an execution strategy object
}
