% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_potential_outcomes.R
\name{expand_conditions}
\alias{expand_conditions}
\title{Expand assignment conditions}
\usage{
expand_conditions(
  formula,
  conditions = c(0, 1),
  assignment_variables = "Z",
  data,
  level = NULL,
  label = NULL
)
}
\arguments{
\item{conditions}{the conditions}

\item{assignment_variables}{the name of assignment variables, if conditions is not already named.}
}
\value{
a data.frame of potential outcome conditions
}
\description{
Internal helper to eagerly build assignment conditions for potential outcomes.
}
\details{
If conditions is a data.frame, it is returned unchanged

Otherwise, if conditions is a list, it is passed to expand.grid for expansion to a data.frame

Otherwise, if condition is something else, box it in a list with assignment_variables for names,
and pass that to expand.grid.
}
\keyword{internal}
