% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pre_register.R
\name{export.pre_registration}
\alias{export.pre_registration}
\title{Export pre-registration document as a PDF, .Rmd, or .Rnw}
\usage{
export.pre_registration(design, file, type = "rmarkdown",
  compile_PDF = TRUE, open_pdf = TRUE)
}
\arguments{
\item{design}{A design object from the declare_design function.}

\item{file}{File name where object is saved in the current working directory.}

\item{type}{Type of document that is created, either \code{knitr} or \code{rmarkdown}, the default.}

\item{compile_PDF}{Indicator for whether code for registration document is compiled into a PDF}

\item{open_PDF}{Indicator for whether PDF is opened in the default PDF viewer after it is compiled.}
}
\value{
Filename and location where .Rmd or .Rnw and PDF file are saved.
}
\description{
Export pre-registration document as a PDF, .Rmd, or .Rnw
}

