% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_estimator.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(data, estimator)
}
\arguments{
\item{data}{A data.frame object, used to fit the model.}

\item{estimator}{An estimator object created by \code{declare_estimator}.}
}
\description{
Fit model
}

