% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/balance.R
\name{get_balance}
\alias{get_balance}
\title{Calculate Covariate Balance}
\usage{
get_balance(covariates, treatment_assignment = "Z", design, data,
  report_difference = FALSE, na.rm = TRUE)
}
\arguments{
\item{covariates}{A character vector of covariate names to be used in the balance assessment.}

\item{treatment_assignment}{The name of the treatment assignment variable in data}

\item{design}{A design object, created by \code{\link{declare_design}}.}

\item{data}{A dataframe, often created by \code{\link{make_data}}.}
}
\value{
A balance object
}
\description{
Description
}
\examples{
examples go here.
}

