% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design.R
\name{get_design}
\alias{get_design}
\title{Get Design from the DeclareDesign Library of Research Designs}
\usage{
get_design(name)
}
\arguments{
\item{name}{Name of the design}
}
\value{
A design object or template function
}
\description{
Get Design from the DeclareDesign Library of Research Designs
}

