% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/design_probabilites.R
\name{get_design_probs}
\alias{get_design_probs}
\title{Calculate probabilties of assignment}
\usage{
get_design_probs(design, data)
}
\arguments{
\item{design}{A design object created by \code{\link{declare_assignment}}; or a function that assigns treatment}

\item{data}{A dataframe, often created by \code{\link{draw_population}} or \code{\link{draw_sample}}.}
}
\value{
A matrix of probabilities of assignment to treatment.
}
\description{
Description
}
\examples{
# these examples don't work yet
# smp <- declare_population(N = 850)
# po <- declare_potential_outcomes(condition_names = c("Z0","Z1"),
#                                    outcome_formula = Y ~ .01 + 0*Z0 + .2*Z1)
# design <- declare_assignment(potential_outcomes = po, m=200)
# mock          <- draw_population(potential_outcomes = po, sample =  smp)
# mock$Z        <- assign_treatment(design, data = mock)
# design_probs <- get_design_probs(design, mock)

head(design_probs)
}

