% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_diagnosand.R
\name{get_diagnosand}
\alias{get_diagnosand}
\title{Get Diagnosand Values}
\usage{
get_diagnosand(diagnosand, simulations)
}
\arguments{
\item{diagnosand}{A \code{diagnosand} object created by \code{\link{declare_diagnosand}}.}

\item{simulations}{A data frame of simulations, typically created by \code{\link{diagnose_design}}.}

\item{na.action}{How the diagnosand summary of diagnostic statistics handles NAs. Can be \link{na.pass}, \link{na.omit}, etc.}
}
\value{
A data.frame of diagnosands.
}
\description{
Get Diagnosand Values
}

