% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_diagnosand.R
\name{get_diagnostic_statistics}
\alias{get_diagnostic_statistics}
\title{Get Diagnostic Statistics}
\usage{
get_diagnostic_statistics(diagnosand, simulations)
}
\arguments{
\item{diagnosand}{A \code{diagnosand} object created by \code{\link{declare_diagnosand}}.}

\item{simulations}{A data frame of simulations, typically created by \code{\link{diagnose_design}}.}
}
\value{
A data.frame of simulations with diagnostic statistics.
}
\description{
Get Diagnostic Statistics
}

