% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{get_estimands}
\alias{get_estimands}
\title{Get Estimands from a Design}
\usage{
get_estimands(design)
}
\arguments{
\item{design}{A design created by \code{\link{declare_design}}.}
}
\value{
a \code{data.frame} of estimands.
}
\description{
Get Estimands from a Design
}
