% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_functions}
\alias{get_functions}
\alias{get_data}
\alias{get_estimates}
\alias{get_assignment}
\alias{get_sample}
\title{Get data, estimates, estimands, assignment vectors, or samples from a design given data}
\usage{
get_data(design, data = NULL, start = 1, end = length(design))

get_estimates(design, data = NULL, start = 1, end = length(design))

get_assignment(design, data = NULL, start = 1, end = length(design))

get_sample(design, data = NULL, start = 1, end = length(design))
}
\arguments{
\item{design}{A design object, typically created using the + operator}

\item{data}{A data.frame object with sufficient information to get the data, estimates, estimands, an assignment vector, or a sample.}

\item{start}{(Defaults to 1) a scalar indicating which step in the design to begin with. By default all data steps are drawn, from step 1 to the last step of the design.}

\item{end}{(Defaults to \code{length(design)}) a scalar indicating which step in the design to finish with.}
}
\description{
Get data, estimates, estimands, assignment vectors, or samples from a design given data
}
