% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_treatment.R
\name{get_observed_assignment_probabilities}
\alias{get_observed_assignment_probabilities}
\title{Reveal probabilties of assignment to realized treatment conditions}
\usage{
get_observed_assignment_probabilities(data, assignment_variable_name,
  assignment)
}
\arguments{
\item{data}{A dataframe, often created by \code{\link{draw_population}} or \code{\link{draw_sample}}.}

\item{assignment_variable_name}{The name of the treatment assignment variable in data.}

\item{assignment}{A assignment object created by \code{\link{declare_assignment}}; or a function that assigns treatment}
}
\value{
A vector probabilities of assignment to treatment.
}
\description{
Description
}
\examples{
population <- declare_population(noise = declare_variable(), N = 1000)
sampling <- declare_sampling(n = 500)
potential_outcomes <- declare_potential_outcomes(formula = Y ~ 5 + .5*Z + noise,
                                                 condition_names = c(0, 1),
                                                 assignment_variable_name = "Z")
assignment <- declare_assignment(condition_names = c(0,1))
pop_draw <- draw_population(population = population)
smp_draw <- draw_sample(data = pop_draw, sampling = sampling)
smp_draw$Z <- assign_treatment_indicator(data = smp_draw, assignment=assignment)

probs <- get_observed_assignment_probabilities(data = smp_draw, 
                        assignment_variable_name= "Z",
                        assignment=assignment)
                        
table(probs)                         

}

