% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_power.R
\name{get_power}
\alias{get_power}
\title{Declare the data-generating process of a variable}
\usage{
get_power(data = NULL, potential_outcomes = NULL, covariates = NULL,
  design, analysis, sims = 100, ...)
}
\arguments{
\item{data}{data object}

\item{design}{design object}

\item{analysis}{analysis object}

\item{sims}{number of iterations}
}
\description{
Declare the data-generating process of a variable
}

