% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_functions.R
\name{get_regression_coefficient}
\alias{get_regression_coefficient}
\title{Extract Regression Coefficients}
\usage{
get_regression_coefficient(model, formula = NULL, coefficient_name,
  statistics = c("est", "se", "p", "ci_lower", "ci_upper", "df"),
  estimates_labels = "")
}
\arguments{
\item{model}{}

\item{formula}{}

\item{coefficient_name}{}

\item{estimates_labels}{}
}
\description{
Extract Regression Coefficients
}

