% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_functions.R
\name{get_regression_coefficient_clustered}
\alias{get_regression_coefficient_clustered}
\title{Extract Regression Coefficients}
\usage{
get_regression_coefficient_clustered(model, formula = NULL, data,
  coefficient_name, cluster_name, label = coefficient_name)
}
\arguments{
\item{model}{An lm model fit}

\item{formula}{An lm formula of the form Y~X}

\item{data}{A dataframe}

\item{coefficient_name}{The name of the coefficient to extract}

\item{cluster_name}{The name of the cluster identifier}

\item{label}{The label for the regression coefficient}
}
\description{
Extract Regression Coefficients
}
\keyword{Cluster,}
\keyword{Robust}

