% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design.R
\name{get_template}
\alias{get_template}
\title{Get Design Template from the DeclareDesign Library of Research Designs}
\usage{
get_template(name)
}
\arguments{
\item{name}{Name of the design template}
}
\value{
An R function
}
\description{
Get Design Template from the DeclareDesign Library of Research Designs
}

