% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_DGP_functions.R
\name{get_variable}
\alias{get_variable}
\title{A function for grabbing variables at a given level in multi-level structures}
\usage{
get_variable(level_ID, variable_name, data, aggregate_function = NULL, ...)
}
\arguments{
\item{level_ID}{The name of the level ID to get a variable for.}

\item{variable_name}{The variable name extracted from the data.}

\item{data}{A data.frame object used to get a variable.}

\item{aggregate_function}{The function to use to aggregate data if applicable.}

\item{...}{A set of additional options sent to the aggregate function.}
}
\description{
A function for grabbing variables at a given level in multi-level structures
}

