% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_DGP_functions.R
\name{get_variable_types}
\alias{get_variable_types}
\title{Print the different kinds of variable types}
\usage{
get_variable_types()
}
\description{
Print the different kinds of variable types
}

