% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{handler_identification}
\alias{handler_identification}
\title{a <- function(x) x
b <- function(x) -x}
\usage{
handler_identification(default_expr, actual_expr)
}
\description{
f <- function(handler = a, x) {
 # Capture expressions from *inside* f()
 handler_names <- handler_identification(quote(a), substitute(handler))
 
 print(handler_names)
 
 handler(x)
 
}
}
\details{
f(b, x)
}
