% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_outcomes.R
\name{has_potential_outcomes}
\alias{has_potential_outcomes}
\title{Determine whether potential outcomes exist in a data frame}
\usage{
has_potential_outcomes(data, potential_outcomes, condition_names = NULL,
  noncompliance = NULL, attrition = NULL)
}
\arguments{
\item{data}{data.frame input}

\item{potential_outcomes}{A potential_outcomes object created by \code{\link{declare_potential_outcomes}}.}

\item{condition_names}{A vector indicating the names of conditions}

\item{noncompliance}{A noncompliance object created by \code{\link{declare_noncompliance}}.}

\item{attrition}{An attrition object created by \code{\link{declare_attrition}}.}
}
\value{
indicator for whether potential outcomes exist
}
\description{
Determine whether potential outcomes exist in a data frame
}

