% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_design.R
\name{implement_design}
\alias{implement_design}
\title{Implement a declared sampling strategy and assignment procedure}
\usage{
implement_design(data, design)
}
\arguments{
\item{data}{}

\item{design}{}
}
\value{
data.frame of the sample if there is a sampling strategy declared and with assigned treatment vector, if defined
}
\description{
Implement a declared sampling strategy and assignment procedure
}

