% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_arm_template.R
\name{make_block_noise}
\alias{make_block_noise}
\title{M-arm Template}
\usage{
make_block_noise(block_var)
}
\arguments{
\item{N}{A scalar (must be an integer) indicating the total population size.}

\item{n}{A scalar (must be an integer) indicating the total sample size. By default, n = N.}

\item{m}{A scalar (must be an integer) indicating the number of arms. Note that m > 1 for any experiment.}

\item{probability_each}{A vector of length m, indicating the probability of assignment to each arm. The first element indicates the control group and subsequent elements indicate each of the treatment groups in ascending order (i.e. treatment 1 to treatment m-1). The elements in the vector must sum to 1. By default, units have a 1/m probability of assignment to each experimental group.}

\item{mu_Y0}{A scalar indicating the expectation of the untreated potential outcome (control group).}

\item{ATEs}{A vector of length m-1 indicating the desired average treatment effect ATE for each experimental group relative to control (mu_Y0).}

\item{noise_scale}{A scalar or a vector of length m indicating the standard deviation of individual-level noise for each experimental group. If it is a scalar, the standard deviation of individual-level noise is constant across all treatment groups. If it is a vector, the standard deviation of noise is group-specific. The elements in the vector correspond to control and then the treatment groups in ascending order.}

\item{coef_X}{A scalar coefficient on a covariate, X, in the potential outcomes function.  By default, coef_X = 0.}

\item{location_scale_X}{A vector of length two indicating the mean and standard deviation of covariate X. By default, location_scale_X = 0.}

\item{cov_adjustment}{A logical argument indicating whether covariate adjustment should be used in estimation. By default, cov_adjustment = F.}

\item{block_var_vector}{For non-clustered designs, a vector indicating the block to which a unit belongs. For cluster-randomized experiments, the vector should be at the cluster level (of length n_clust_pop).}

\item{block_var_probs}{For non-clustered designs, a vector specifying the distribution of a multinomial blocking variable (probability per level). The elements in the vector must be between 0 and 1 and sum to 1.}

\item{blocked_RA}{A logical argument indicating whether blocked random assignment should be used. By default, blocked_RA = F.}

\item{block_prob_each}{A matrix with the same number of rows as blocks and the same number of columns as treatment arms. Cell entries are the probabilites of assignment to treatment within each block. The rows should respect the ordering of the blocks as determined by sort(unique(block_var)). Use only if the probabilities of assignment should vary by block, otherwise use prob_each. Each row of block_prob_each must sum to 1.}

\item{n_clust_pop}{A scalar (must be an integer) indicating the number of equal-sized clusters in the population.}

\item{n_clust_samp}{A scalar (must be an integer) indicating the number of equal-sized clusters in the sample.}

\item{n_blocks}{For clustered designs, users can specify the number of equal-sized blocks in the population.}

\item{SE}{Indicates the standard error estimator. Options include "OLS" for classical OLS standard errors, "robust" for Huber-White heteroskedasticity-robust standard errors, or "clustered" for robust clustered standard errors. By default, the "robust" standard errors are used for non-clustered designs and "clustered" standard errors are used for clustered designs.}

\item{custom_pop_list}{Allows users to input cluster-assigned designs with clusters and blocks of varying sizes. Users should input a list of three vectors: (1) individuals using rep(1, N), where N is the population size; (2), clusters, a vector containing the number per cluster such that the sum of the vector is N; and (3) blocks, a vector containing the number of clusters in each block such that the sum of the vector is equal to the length of the cluster vector. All elements must be positive integers.}
}
\value{
A design
}
\description{
M-arm Template
}

