% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_covariates.R
\name{make_covariates}
\alias{make_covariates}
\title{Create covariates for the experiment}
\usage{
make_covariates(..., design_object = NULL, N = design_object$N)
}
\arguments{
\item{...}{A list of DGP_objects (of the form variable = declare_DGP(...)), or functions that generate random numbers, of the form variable = function()rng(...)}

\item{design_object}{The object created by declare_design}

\item{N}{The number of observations.}
}
\description{
Create covariates for the experiment
}

