% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_data_frame.R
\name{make_data}
\alias{make_data}
\title{Construct data frame for mock analysis using potential outcomes and covariates}
\usage{
make_data(potential_outcomes, covariates)
}
\arguments{
\item{potential_outcomes}{tmp}

\item{covariates}{tmp}

\item{...}{additional options to be sent to the analysis function and the test_success function}
}
\value{
a data frame
}
\description{
Description
}
\examples{
# these examples don't work yet
# declare_analysis(analysis = "diff-in-means")
# declare_analysis(analysis = function(Y, Z, data) lm(paste(Y, "~", Z), data = data))
}

