% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_design.R
\name{merge_estimates_inquiries}
\alias{merge_estimates_inquiries}
\title{Join estimates to their inquiries}
\usage{
merge_estimates_inquiries(estimates, inquiries)
}
\arguments{
\item{estimates}{A tibble of estimator output.}

\item{inquiries}{A tibble of inquiry output.}
}
\value{
A tibble.
}
\description{
Used in \code{\link[=draw_estimates]{draw_estimates()}} and \code{\link[=simulate_design]{simulate_design()}} to attach estimands to
estimator output via the \code{inquiry} label.
}
\keyword{internal}
