% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{modify_design}
\alias{modify_design}
\alias{add_step}
\alias{remove_step}
\alias{replace_step}
\title{Modify a design object}
\usage{
modify_design(design, ...)

add_step(..., before = NULL, after = NULL)

remove_step(...)

replace_step(..., replace)
}
\arguments{
\item{design}{a design object, usually created by \code{\link{declare_design}}, \code{\link{quick_design}}, or \code{\link{download_design}}.}

\item{...}{a series of calls to \code{\link{add_step}}, \code{\link{remove_step}}, or \code{\link{replace_step}}}
}
\value{
A design object. See \code{\link{declare_design}} for details.
}
\description{
Modify a design object
}
\examples{





}
