% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reveal_design.R
\name{observed_outcome}
\alias{observed_outcome}
\title{Reveal observed outcomes based on a given treatment assignment}
\usage{
observed_outcome(outcome = "Y", treatment_assignment, data, sep = "_")
}
\arguments{
\item{outcome}{A character string}

\item{treatment_assignment}{A string indicating the name of the realized treatment assignment vector}

\item{data}{A data frame including the outcome and realized treatment assignment vectors indicating by outcome and treatment_assignment}

\item{sep}{The character separating outcomes from condition names in the potential outcomes columns of data}
}
\value{
an outcome vector of observed y
}
\description{
Description
}
\examples{
smp <- declare_population(N = 850)
po <- declare_potential_outcomes(condition_names = c("Z0","Z1"),
                                   outcome_formula = Y ~ .01 + 0*Z0 + .2*Z1)
design <- declare_design(potential_outcomes = po, m=200)
mock          <- draw_population(potential_outcomes = po, sample =  smp)
mock$Z        <- assign_treatment(design, data = mock)
mock$Y  <- observed_outcome("Y", "Z", mock)
summary(lm(Y~Z, data=mock))
}

