% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pre_register_paper_draft.R
\name{paper_draft}
\alias{paper_draft}
\title{Create paper draft}
\usage{
paper_draft(design, covariates, potential_outcomes, analysis, title, authors,
  affiliations, acknowledgements, abstract, random.seed = 42, dir = getwd(),
  temp_dir = FALSE, format = "rmarkdown", make_output = TRUE,
  keep_tex = FALSE, save_r_code = FALSE, output_format = "pdf",
  open_output = TRUE, ...)
}
\arguments{
\item{design}{A design object from the declare_design function.}

\item{analysis}{An analysis object from the declare_analysis function, or a list of analysis objects.}

\item{random.seed}{Random seed to ensure reproducibility of the design.}

\item{make_output}{Indicator for whether code for registration document is compiled into a PDF, Microsoft Word, or HTML document}

\item{output_format}{String indicating which type of output file is created, "pdf" (default), "word", or "html"}

\item{open_output}{Indicator for whether the output file is opened after it is compiled.}

\item{...}{Other options for the render() command to create the output file from the markdown code.}

\item{data}{A data object from the make_y function.}

\item{registration_title}{Title of the pre-registration.}

\item{registration_authors}{List of authors, consisting of vectors of two strings -- one for the author's name and one for her affiliation. e.g. \code{list(c("Alan Gerber", "Yale University"), c("Donald Green", "Columbia University"))}. Optionally, a third item in the vector can be the text for a footnote with author contact information.}

\item{registration_abstract}{General description of the experiment.}

\item{file}{File name where object is saved.}

\item{type}{Type of document that is created, either \code{knitr} or \code{rmarkdown}, the default.}

\item{check_registration}{Indicates whether the design is evaluated for consistency with declared analyses.}
}
\value{
Filename and location where .Rmd or .Rnw and PDF file are saved.
}
\description{
Create paper draft
}

