% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot.design}
\alias{plot.design}
\title{Plot a flowchart of a design}
\usage{
\method{plot}{design}(x, ...)
}
\arguments{
\item{x}{An object of class `design`, typically created by}

\item{...}{Plotting and visualization arguments; currently supported `main`,
`theme`.}
}
\description{
This method plots objects generated in DeclareDesign.
}
