% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diagnosis.R
\name{plot.diagnosis}
\alias{plot.diagnosis}
\title{Default plot method for diagnosis object}
\usage{
\method{plot}{diagnosis}(x, type = "coverage", ...)
}
\arguments{
\item{x}{Design diagnosis object created by \link{diagnose_design}.}

\item{type}{Type of the plot to draw. Currently only supports \code{type = "coverage"},}

\item{...}{Additional parameters passed to ggplot \code{_facet} function.}
}
\value{
Diagnosis plot produced by ggplot2.
}
\description{
Default plot method for diagnosis object
}

