% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_outcomes_and_covariates.R
\name{plot_covariates}
\alias{plot_covariates}
\title{Plot the frequency or density of covariates}
\usage{
plot_covariates(data, outcome_names, color_palette = NULL, barplot = FALSE,
  density = FALSE, histogram = FALSE, ...)
}
\arguments{
\item{data}{What is it?}

\item{outcome_names}{What is it?}

\item{color_palette}{What is it?}

\item{barplot}{What is it?}

\item{density}{What is it?}

\item{histogram}{What is it?}

\item{...}{What is it?}
}
\description{
Plot the frequency or density of covariates
}

