% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_outcomes_and_covariates.R
\name{plot_outcomes}
\alias{plot_outcomes}
\title{Plot the frequency or density of potential outcomes}
\usage{
plot_outcomes(data, outcome_names, color_palette = NULL, barplot = FALSE,
  density = FALSE, histogram = FALSE, ...)
}
\arguments{
\item{put}{in params}
}
\description{
Plot the frequency or density of potential outcomes
}

