% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_power.R
\name{plot_power}
\alias{plot_power}
\title{Plot power across values of a parameter like the sample size}
\usage{
plot_power(data, design, analysis, vary_parameter = "N", vary_sequence)
}
\arguments{
\item{data}{data object}

\item{design}{design object}

\item{analysis}{analysis object}

\item{vary_parameter}{indicating which parameter is varied, default is "N"}

\item{vary_sequence}{set of values of the parameter to calculate power for}

\item{sims}{number of iterations}
}
\description{
Plot power across values of a parameter like the sample size
}

