% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{+.d_par}
\alias{+.d_par}
\title{Declare design via the `+` operator}
\usage{
\method{+}{d_par}(lhs, rhs)
}
\arguments{
\item{lhs}{An object of class \code{dd} or \code{design}.}

\item{rhs}{An object of class \code{dd} or \code{design}.}
}
\value{
A \code{design} object.
}
\description{
Declare design via the `+` operator
}
\examples{

pop <- declare_population(N = 100)
smp <- declare_sampling(n = 50)
assgn <- declare_assignment(m = 25)

my_design <- pop + smp + assgn

}
