% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R, R/get_functions.R
\name{post_design}
\alias{post_design}
\alias{print_code}
\alias{get_data}
\alias{get_estimates}
\alias{get_assignment}
\alias{get_sample}
\title{Explore your design}
\usage{
print_code(design)

get_data(design, data = NULL, start = 1, end = length(design))

get_estimates(design, data = NULL, start = 1, end = length(design))

get_assignment(design, data = NULL, start = 1, end = length(design))

get_sample(design, data = NULL, start = 1, end = length(design))
}
\arguments{
\item{design}{A design object, typically created using the + operator}

\item{data}{A data.frame object with sufficient information to conduct data generation from the start step to the end step.}

\item{start}{(Defaults to 1) a scalar indicating which step in the design to begin drawing data from. By default all data steps are drawn, from step 1 to the last step of the design.}

\item{end}{(Defaults to \code{length(design)}) a scalar indicating which step in the design to finish drawing data by.}

\item{design}{A design object, typically created using the + operator}

\item{data}{A data.frame object with sufficient information to run estimators.}

\item{start}{(Defaults to 1) a scalar indicating which step in the design to begin getting estimates from. By default all estimators are calculated, from step 1 to the last step of the design.}

\item{end}{(Defaults to \code{length(design)}) a scalar indicating which step in the design to finish getting estimates from.}

\item{data}{A data.frame object with sufficient information to draw assignment vectors.}

\item{start}{(Defaults to 1) a scalar indicating which step in the design to begin drawing assignment vectors. By default all assignment vectors are drawn, from step 1 to the last step of the design.}

\item{end}{(Defaults to \code{length(design)}) a scalar indicating which step in the design to finish drawing assignment vectors.}

\item{data}{A data.frame object with sufficient information to run sampling steps.}

\item{start}{(Defaults to 1) a scalar indicating which step in the design to begin sampling. By default all sampling steps are run, from step 1 to the last step of the design.}

\item{end}{(Defaults to \code{length(design)}) a scalar indicating which step in the design to finish sampling.}
}
\description{
Explore your design

Print code to recreate a design
}
\examples{

design <-
  declare_population(N = 500, noise = rnorm(N)) +
  declare_potential_outcomes(Y ~ noise + Z * rnorm(N, 2, 2)) +
  declare_sampling(n = 250) +
  declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_step(dplyr::mutate, noise_sq = noise^2) +
  declare_assignment(m = 25) +
  declare_reveal() +
  declare_estimator(Y ~ Z, estimand = "my_estimand")

design

df <- draw_data(design)

estimates <- draw_estimates(design)
estimands <- draw_estimands(design)


my_population <- declare_population(N = 100)

my_assignment <- declare_assignment(m = 50)

my_design <- my_population + my_assignment

print_code(my_design)

}
