% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pre_register_paper_draft.R
\name{pre_register}
\alias{pre_register}
\title{Create pre-registration document}
\usage{
pre_register(design, clusters = NULL, blocks = NULL, sample_frame = NULL,
  potential_outcomes = NULL, analysis = NULL, data = NULL, title = NULL,
  authors = NULL, affiliations = NULL, acknowledgements = NULL,
  abstract = NULL, random_seed = 42, dir = getwd(), temp_dir = FALSE,
  format = "rmarkdown", check_registration = TRUE, make_output = TRUE,
  keep_tex = FALSE, save_r_code = FALSE, output_format = "pdf",
  open_output = TRUE)
}
\arguments{
\item{design}{A design object from the declare_design function.}

\item{clusters}{what is it?}

\item{blocks}{what is it?}

\item{sample_frame}{what is it?}

\item{potential_outcomes}{what is it?}

\item{analysis}{what is it?}

\item{data}{A data object from the make_data function.}

\item{title}{Title of the pre-registration.}

\item{authors}{List of authors, consisting of vectors of two strings -- one for the author's name and one for her affiliation. e.g. \code{list(c("Alan Gerber", "Yale University"), c("Donald Green", "Columbia University"))}. Optionally, a third item in the vector can be the text for a footnote with author contact information.}

\item{affiliations}{what is it?}

\item{acknowledgements}{what is it?}

\item{abstract}{General description of the experiment.}

\item{random_seed}{Random seed to ensure reproducibility of the design.}

\item{dir}{what is it?}

\item{temp_dir}{what is it?}

\item{format}{what is it?}

\item{check_registration}{what is it?}

\item{make_output}{Indicator for whether code for registration document is compiled into a PDF, Microsoft Word, or HTML document}

\item{keep_tex}{what is it?}

\item{save_r_code}{what is it?}

\item{output_format}{String indicating which type of output file is created, "pdf" (default), "word", or "html"}

\item{open_output}{Indicator for whether the output file is opened after it is compiled.}
}
\value{
what does it return?
}
\description{
Description
}

