% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_diagnoses.R
\name{print.compared_diagnoses}
\alias{print.compared_diagnoses}
\title{Print a diagnosand comparison}
\usage{
\method{print}{compared_diagnoses}(x, ...)
}
\arguments{
\item{x}{(required) A \code{compared_diagnoses} object, as returned by
\code{\link[=compare_diagnoses]{compare_diagnoses()}}.}

\item{...}{(optional) Passed to the \code{\link[=print]{print()}} method of
\code{compared_diagnoses_df}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a diagnosand comparison
}
\examples{
design <- declare_parameters(n_units = 100) +
  declare_model(N = n_units, Y = rnorm(N), Z = randomizr::complete_ra(N)) +
  declare_inquiry(ATE = 0) +
  declare_estimator(Y ~ Z, .method = lm, term = "Z", inquiry = "ATE",
                    label = "ols")
print(compare_diagnoses(design, redesign(design, n_units = 200),
                        sims = 20, bootstrap_sims = 20))
}
