% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs_methods.R
\name{print.design_comparison}
\alias{print.design_comparison}
\title{print.design_comparison
Print function for design comparison.}
\usage{
\method{print}{design_comparison}(design_comparison, highlights = TRUE)
}
\arguments{
\item{design_comparison}{A comparison generated by compare_designs.}

\item{highlights}{Display highlights (as opposed to full comparison)?}
}
\description{
print.design_comparison
Print function for design comparison.
}
