% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.design_step}
\alias{print.design_step}
\title{Print a design step}
\usage{
\method{print}{design_step}(x, ...)
}
\arguments{
\item{x}{(required) A \code{design_step}.}

\item{...}{(optional) Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a design step
}
\examples{
s <- declare_model(N = 10, Y = rnorm(N))
print(s)
}
