% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosands.R
\name{print.diagnosands}
\alias{print.diagnosands}
\title{Print a diagnosands object}
\usage{
\method{print}{diagnosands}(x, ...)
}
\arguments{
\item{x}{A \code{diagnosands} object.}

\item{...}{Ignored.}
}
\value{
The input invisibly.
}
\description{
Print a diagnosands object
}
\examples{
print(default_diagnosands())
}
