% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{print_code}
\alias{print_code}
\title{Print code to recreate a design}
\usage{
print_code(design)
}
\arguments{
\item{design}{A design object, typically created using the + operator}
}
\description{
Print code to recreate a design
}
\examples{

my_population <- declare_population(N = 100)

my_assignment <- declare_assignment(m = 50)

my_design <- my_population + my_assignment

print_code(my_design)

}
