% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_functions.R
\name{proportion_potential_outcomes_function}
\alias{proportion_potential_outcomes_function}
\title{Default function for proportional potential outcomes}
\usage{
proportion_potential_outcomes_function(data, population_proportions = c(0.5,
  0.5), condition_names = NULL, assignment_variable_name = "Z")
}
\arguments{
\item{data}{A data.frame}

\item{population_proportions}{Proportion in outcome for each treatment condition.}

\item{condition_names}{A vector of condition names.}

\item{assignment_variable_name}{The variable name for the assignment indicator.}
}
\description{
Default function for proportional potential outcomes
}

