% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_template.R
\name{rdd_template}
\alias{rdd_template}
\title{Template to generate regression discontinuity designs}
\usage{
rdd_template(N = 100, formula = y ~ running + Z + Z * running,
  running_variable = declare_variable(type = "normal"),
  baseline_variable = declare_variable(type = "normal", location_scale = c(0,
  0.1)), polynomial = 4, control_coefs = c(1, 0, 0, 0),
  treatment_coefs = c(0.02, 0.02, 0.02, 0.2), cutoff = NULL,
  bandwidth = NULL, plot_responses = TRUE, coefficient_name = "Z")
}
\arguments{
\item{N}{Number of units.}

\item{formula}{The estimating formula.}

\item{running_variable}{A \code{\link{declare_variable}} expression that specifies the distribution of the running variable. Standard normal by default.}

\item{baseline_variable}{A \code{\link{declare_variable}} expression that specifies the distribution of the baseline variable that is added to each unit's potential outcomes. Standard normal by default.}

\item{polynomial}{The number of polynomial terms in the response functions.}

\item{control_coefs}{The coefficients in the polynomial response function of the control.}

\item{cutoff}{The location of the discontinuity. Median of the running variable by default.}

\item{bandwidth}{The size of the window around the cutoff to include in the sample. Includes all units by default.}

\item{coefficient_name}{In case custom formula are specified, the coefficient that should be extracted from the model fit as an estimate.}

\item{control_coefs}{The coefficients in the polynomial response function of the treatment.}

\item{plot_response}{Logical indicator for whether to plot a representation of the response functions.}
}
\value{
A design
}
\details{
This templating function generates a sharp regression discontinuity design based on continuity assumptions.
}

