% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{rename_dots}
\alias{rename_dots}
\title{Rename dots to match a handler’s formals}
\usage{
rename_dots(handler, dots)
}
\arguments{
\item{handler}{A function whose formals to match.}

\item{dots}{A list of quosures/args.}
}
\value{
The \code{dots} list with names adjusted (may add a \code{data} symbol).
}
\description{
Given a function and dots, rename dots based on how things will positionally match
}
\examples{
DeclareDesign:::rename_dots(fabricatr::fabricate, list(N = 3, X = 1))
}
\keyword{internal}
