% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_analysis.R
\name{run_analysis.analysis}
\alias{run_analysis.analysis}
\title{Runs a pre-defined experimental analysis}
\usage{
run_analysis.analysis(analysis, data)
}
\arguments{
\item{analysis}{analysis object created by declare_analysis}

\item{data}{data object created by make_y}

\item{Y}{name of the outcome for the analysis, a character object}

\item{Z}{name of the treatment indicator(s) for the analysis, a character object or vector of character objects}
}
\value{
a numeric scalar or vector of p-values
}
\description{
Description
}
\examples{
##Some examples will go here
}

