% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{run_simulations}
\alias{run_simulations}
\title{Run simulations of a design's frequentist properties}
\usage{
run_simulations(design = NULL, sims = 5)
}
\arguments{
\item{sims}{number of iterations}

\item{population}{A population object, made with \code{\link{declare_population}}}

\item{sampling}{A sampling object, made with \code{\link{declare_sampling}}}

\item{label}{label for the simulation}

\item{analysis_labels}{labels for each analysis}
}
\description{
Run simulations of a design's frequentist properties
}

