% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs_methods.R
\name{save_design_comparison}
\alias{save_design_comparison}
\title{save_design_comparison
Save function for design comparison. Renders a document (e.g., pdf).}
\usage{
save_design_comparison(design_comparison, output_type = c("html", "pdf",
  "word", "github", "Rmd"), file_prefix = "design_comparison",
  render_quietly = FALSE)
}
\arguments{
\item{design_comparison}{A comparison generated by compare_designs.}

\item{output_type}{Creates a file called 'design_comparison.html' unless file_prefix is provided (or a different file type is specified). 'Rmd' means write R markdown only (otherwise Rmd is saved along with desired type of output file). Options: c("html", "pdf", "word", "github", "Rmd").}

\item{file_prefix}{Optional. If provided, names the file specified by 'output_type.'}

\item{render_quietly}{If FALSE (default), preview pops up and rendering details display to console.}
}
\description{
save_design_comparison
Save function for design comparison. Renders a document (e.g., pdf).
}
