% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_inputs.R
\name{save_design_replication_files}
\alias{save_design_replication_files}
\title{Save replication files for a design}
\usage{
save_design_replication_files(design)
}
\arguments{
\item{design}{A design object created by \code{\link{declare_design}}.}
}
\description{
Save replication files for a design
}

