% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_environment.R
\name{set_environment}
\alias{set_environment}
\title{Set elements for the environment of a design}
\usage{
set_environment(design, ...)
}
\arguments{
\item{...}{Named objects.}
}
\value{
a design object with parameters set in the global environment
}
\description{
A design accesses elements in the global environment that may not be available on later use. Setting these lements saves them into the design step envionments.
}
\examples{
n = 10
design <- declare_model(N = n) + NULL
draw_data(design) 
design <- set_environment(design, n = 5)
rm(n)
draw_data(design) 

}
