% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_experiment.R
\name{simulate_experiment}
\alias{simulate_experiment}
\title{Declare the data-generating process of a variable}
\usage{
simulate_experiment(data = NULL, potential_outcomes = NULL,
  sample_frame = NULL, blocks = NULL, clusters = NULL, design, analysis,
  bootstrap_data = FALSE, N_bootstrap, sims = 5, label = NULL)
}
\arguments{
\item{data}{data object}

\item{design}{design object}

\item{analysis}{analysis object}

\item{sims}{number of iterations}
}
\description{
Declare the data-generating process of a variable
}

