% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_template.R
\name{spillover_template}
\alias{spillover_template}
\title{Template to generate designs with spillovers}
\usage{
spillover_template(treatment_effect = 1, spillover_effect = 0,
  latitude_effect = 2, buffer = 0)
}
\arguments{
\item{treatment_effect}{A real number that indicates the size of the average treatment effect.}

\item{spillover_effect}{A real number that indicates the size of the spillover effect.}

\item{latitude_effect}{A real number that correlates latitude with potential outcomes. The higher the value, the more correlated potential outcomes are with spatial position.}

\item{buffer}{The minimum units of distance between the two sampled units. 0 by default.}
}
\value{
A design
}
\details{
This templating function generates a design in which two units are sampled from a population of 36 units that are aligned in a grid in space. The units can be sampled such that they are a minimum distance from one another using a spatial buffer: each unit is one unit of measurement distant from its vertical or horizontal neighbor and sqrt(2) from its diagonal neighbors. One unit is assigned to treatment and one to control. The user can specify the average treatment effect, the effect of the spillover, the degree to which spatial position (latitude) is correlated with potential outcomes, and the size of the spatial buffer separating units.
}

