% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosands.R
\name{stock_diagnosand_names}
\alias{stock_diagnosand_names}
\title{The names of the stock diagnosands}
\usage{
stock_diagnosand_names()
}
\value{
A character vector.
}
\description{
The names of the stock diagnosands
}
\examples{
stock_diagnosand_names()
}
