% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs_methods.R
\name{summary.design_comparison}
\alias{summary.design_comparison}
\title{summary.design_comparison
Summary function for design comparison. See also save_design_comparison().}
\usage{
\method{summary}{design_comparison}(design_comparison)
}
\arguments{
\item{design_comparison}{A comparison generated by compare_designs.}
}
\description{
summary.design_comparison
Summary function for design comparison. See also save_design_comparison().
}
