% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.potential_outcomes}
\alias{summary.potential_outcomes}
\title{Summarizing potential outcomes}
\usage{
\method{summary}{potential_outcomes}(object, extended = TRUE, ...)
}
\arguments{
\item{object}{A potential_outcomes object as created by \code{\link{declare_potential_outcomes}}.}

\item{extended}{Logical. Whether to print extended summary including labels of potential outcomes delcared and function call which re-creates the potential outcomes object. Default is \code{TRUE}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
If \code{extended = FALSE} function returns list with one element, \code{summary_text}, which contains text explanation of potential outcomes structure. If \code{extended = TRUE}, function returns list with 3 elements: \code{summary_text} as explained above, \code{po_labels}, which contains a named list of labels for all declared potential outcomes, and \code{code}, which contains function call which creates potential outcomes object supplied.
}
\description{
Summary method for potential_outcomes object created by \code{\link{declare_potential_outcomes}}.
}
\examples{
# ADD EXAMPLES HERE

}

