% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_analysis.R
\name{test_success.analysis}
\alias{test_success.analysis}
\title{Return the result of a test for treatment effect(s) from an experimental analysis}
\usage{
test_success.analysis(formula, analysis, finished_analysis = NULL, design,
  data, alpha = 0.05)
}
\arguments{
\item{analysis}{analysis object created by declare_analysis}

\item{design}{design object created by declare_design}

\item{data}{data object created by make_y}
}
\value{
a numeric scalar or vector of p-values
}
\description{
Description
}
\examples{
# Some examples will go here
}

