% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_filter.R
\name{tidy_filter}
\alias{tidy_filter}
\title{Tidy Model Results and Filter to Relevant Coefficients}
\usage{
tidy_filter(fit, term = FALSE)
}
\arguments{
\item{fit}{A model fit, as returned by a modeling function like lm, glm, or estimatr::lm_robust.}

\item{term}{A character vector of the terms that represent quantities of interest, i.e., "Z". If FALSE, return the first non-intercept term; if TRUE return all terms.}
}
\value{

}
\description{
Tidy Model Results and Filter to Relevant Coefficients
}
