% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_step.R
\name{tidy_step}
\alias{tidy_step}
\title{Convert a step to a tidy function for use in a design}
\usage{
tidy_step(...)
}
\arguments{
\item{...}{A function of data and other arguments that returns data}
}
\value{
A function that takes only data compatible with creating a design
}
\description{
Wrap a data-in, data-out function call as a tidy function
}
\examples{

# With dplyr verbs like mutate
library(dplyr)
my_population <- declare_population(N = 10)
a_wrapped_mutate <- tidy_step(mutate(q = 5))

my_design <- my_population + a_wrapped_mutate
my_design

# With user-written functions

my_function <- function(data, my_mean) {
data$new_variable <- rnorm(n = nrow(data), mean = my_mean)
data
}

my_design <- my_population + tidy_step(my_function(my_mean = 5))
my_design

}
