% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vars}
\alias{vars}
\title{Capture variables as quosures}
\usage{
vars(...)
}
\arguments{
\item{...}{(optional) Expressions, captured unevaluated.}
}
\value{
A list of quosures of class \code{quosures}, one per expression.
}
\description{
Convenience alias for \code{\link[rlang:quos]{rlang::quos()}} used inside design declarations to
pass through bare expressions.
}
\examples{
qs <- vars(x, y, z)
length(qs)
}
