% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_step.R
\name{wrap_step}
\alias{wrap_step}
\title{Wrap a step}
\usage{
wrap_step(...)
}
\arguments{
\item{...}{}
}
\value{
A function that takes only data. (is a quosure).
}
\description{
Wrap a data-in, data-out function call
}
\examples{

# With dplyr verbs like mutate
library(dplyr)
my_population <- declare_population(N = 10)
a_wrapped_mutate <- wrap_step(mutate(q = 5))

my_design <- my_population + a_wrapped_mutate
my_design

# With user-written functions

my_function <- function(data, my_mean) {
data$new_variable <- rnorm(n = nrow(data), mean = my_mean)
data
}

my_design <- my_population + wrap_step(my_function)
my_design


}
