% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{add_step}
\alias{add_step}
\title{Modify a Design by Adding Steps}
\usage{
add_step(..., before = NULL, after = NULL)
}
\arguments{
\item{...}{steps to add to a design}

\item{before}{bare (unquoted) name of the step before which to add steps.}

\item{after}{bare (unquoted) name of the step after which to add steps.}
}
\description{
Modify a Design by Adding Steps
}
\details{
see \code{\link{modify_design}} for details.
}
