% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign_treatment_reveal_obs_y.R
\name{assign_treatment}
\alias{assign_treatment}
\title{Assign treatment status}
\usage{
assign_treatment(design)
}
\arguments{
\item{design}{A design object created by declare_design; or a function that assigns treatment}

\item{...}{Options to override defaults in random assignmetn function}
}
\value{
a random assignment vector
}
\description{
Description
}
\examples{
# some examples go here
}

