% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign_treatment.R
\name{assign_treatment}
\alias{assign_treatment}
\title{Assign treatment}
\usage{
assign_treatment(data, assignment, random_seed = NULL)
}
\arguments{
\item{data}{A data.frame, often created by \code{\link{draw_population}} or \code{\link{draw_sample}}.}

\item{assignment}{A assignment object created by \code{\link{declare_assignment}}; or a function that assigns treatment}

\item{random_seed}{An optional random seed, in order to ensure the replicability of a particular random assignment.}
}
\value{
A data.frame with new columns added for a treatment assignment, probabilities of assignment, and inverse probability weights.
}
\description{
This function takes a data.frame and an assignment object and returns a data.frame with a treatment assignment, probabilities of assignment, and inverse probability weights.
}
\examples{
population <- declare_population(noise = declare_variable(), N = 1000)
sampling <- declare_sampling(n = 500)
potential_outcomes <- declare_potential_outcomes(formula = Y ~ 5 + .5*Z + noise,
                                                 condition_names = c(0, 1),
                                                 assignment_variable_name = "Z")
assignment <- declare_assignment(condition_names = c(0,1))
pop_draw <- draw_population(population = population)
smp_draw <- draw_sample(data = pop_draw, sampling = sampling)
smp_draw <- assign_treatment(data = smp_draw, assignment = assignment)

head(smp_draw)
}

