% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_custom_functions.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Bootstrap data, including multi-level data}
\usage{
bootstrap_data(data, size, level_IDs = NULL)
}
\arguments{
\item{data}{A data.frame, usually provided by the user and not created with the DeclareDesign package.}

\item{size}{A size declaration -- must describe this!}

\item{level_IDs}{The variables that indicate the data hierarchy.}
}
\value{
A data.frame
}
\description{
Bootstrap data, including multi-level data
}

