% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{check_design}
\alias{check_design}
\title{Check Design Object}
\usage{
check_design(design)
}
\arguments{
\item{design}{A design object created by \code{\link{declare_design}}.}
}
\description{
Check consistency in design object
}

