% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_registration.R
\name{check_registration}
\alias{check_registration}
\title{Check a pre-registration for consistency between design and analysis}
\usage{
check_registration(design, sample_frame, potential_outcomes, blocks, clusters,
  analysis, sims)
}
\arguments{
\item{design}{design object}

\item{sample_frame}{covariates object}

\item{potential_outcomes}{potential outcomes object}

\item{blocks}{what is it?}

\item{clusters}{what is it?}

\item{analysis}{analysis object}

\item{sims}{number of iterations}
}
\description{
Check a pre-registration for consistency between design and analysis
}

