% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_diagnosands.R
\name{cite_design}
\alias{cite_design}
\title{Retrieve a design's citation (stub)}
\usage{
cite_design(design)
}
\arguments{
\item{design}{A \code{design}.}
}
\value{
The citation list, or \code{NULL} if none has been set.
}
\description{
Retrieve a design's citation (stub)
}
\examples{
design <- set_citation(declare_model(N = 10, Y = rnorm(N)),
                       title = "Example")
cite_design(design)
}
