% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{cite_design}
\alias{cite_design}
\title{Obtain the preferred citation for a design}
\usage{
cite_design(design, ...)
}
\arguments{
\item{design}{a design object created using the + operator}

\item{...}{options for printing the citation if it is a BibTeX entry}
}
\description{
Obtain the preferred citation for a design
}
