% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{compare_designs}
\alias{compare_designs}
\alias{compare_design_code}
\alias{compare_design_data}
\alias{compare_design_estimates}
\alias{compare_design_inquiries}
\alias{compare_design_summaries}
\alias{compare_diagnoses}
\alias{print_code}
\title{Defunct comparison helpers}
\usage{
compare_designs(...)

compare_design_code(...)

compare_design_data(...)

compare_design_estimates(...)

compare_design_inquiries(...)

compare_design_summaries(...)

compare_diagnoses(...)

print_code(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
Never returns.
}
\description{
These functions existed in the original DeclareDesign but are not part of
DeclareDesignZero. They error with a message asking the user to share the
R script that defines the design.
}
