% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_designs}
\alias{compare_designs}
\title{Compare the Properties of Several Research Designs}
\usage{
compare_designs(..., sims = 500)
}
\arguments{
\item{...}{A set of designs created by \code{\link{declare_design}}.}

\item{sims}{The number of simulations, defaulting to 500.}
}
\description{
Compare the Properties of Several Research Designs
}
