% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_designs}
\alias{compare_designs}
\title{Compare designs}
\usage{
compare_designs(design, population = NULL, sampling = NULL,
  assignment = NULL, estimator = NULL, potential_outcomes = NULL,
  inputs = NULL, size = NULL, labels = NULL, population_draws = 5,
  sample_draws = 5)
}
\arguments{
\item{design}{A design object created by \code{\link{declare_design}}.}

\item{population}{A population object created by \code{\link{declare_population}}.}

\item{sampling}{A sampling object created by \code{\link{declare_sampling}}.}

\item{assignment}{An assignment object created by \code{\link{declare_assignment}}.}

\item{estimator}{A estimator object created by \code{\link{declare_population}}.}

\item{potential_outcomes}{A potential_outcomes object created by \code{\link{declare_population}}.}

\item{inputs}{An inputs object created by \code{\link{declare_population}}.}

\item{size}{A vector representing the size of the population.}

\item{labels}{A vector of labels for each design that will be compared.}

\item{population_draws}{A number of draws of the population.}

\item{sample_draws}{A number of draws of the sample for each draw of the population.}
}
\description{
Create a diagnosis of two or more designs
}

