% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_designs}
\alias{compare_designs}
\title{Compare designs}
\usage{
compare_designs(design, population = NULL, sampling = NULL,
  assignment = NULL, estimator = NULL, potential_outcomes = NULL,
  inputs = NULL, size = NULL, labels = NULL, population_draws = 5,
  sample_draws = 5)
}
\arguments{
\item{design}{}

\item{population}{}

\item{sampling}{}

\item{assignment}{}

\item{estimator}{}

\item{potential_outcomes}{}

\item{inputs}{}

\item{size}{}

\item{labels}{}

\item{population_draws}{}

\item{sample_draws}{}
}
\description{
Create a diagnosis of two or more designs
}

