% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_diagnoses.R
\name{compare_diagnoses_internal}
\alias{compare_diagnoses_internal}
\title{Internal method to compare diagnoses}
\usage{
compare_diagnoses_internal(diagnosis1, diagnosis2, sims, bootstrap_sims,
  merge_by_estimator)
}
\arguments{
\item{diagnosis1}{A diagnosis.}

\item{diagnosis2}{A diagnosis.}

\item{merge_by_estimator}{A logical. Whether to include \code{estimator_label} in the set of columns used for merging. Defaults to \code{TRUE}}
}
\value{
A list with a data frame of compared diagnoses and both diagnoses.
}
\description{
Internal method to compare diagnoses
}
\keyword{internal}
