% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare_experiments.R
\name{compare_experiments}
\alias{compare_experiments}
\title{Declare the data-generating process of a variable}
\usage{
compare_experiments(N = NULL, N_clusters = NULL, N_per_cluster = NULL,
  design = NULL, analysis = NULL, sample_frame = NULL,
  potential_outcomes = NULL, blocks = NULL, clusters = NULL, sims = 5)
}
\arguments{
\item{design}{design object}

\item{analysis}{analysis object}

\item{sims}{number of iterations}

\item{data}{data object}
}
\description{
Declare the data-generating process of a variable
}
\examples{
## imagined options:
  ## compare_experiments(N = c(500, 1000, 2000), m = 5, design = design, etc.)
  ## compare_experiments(N = c(500, 1000, 2000), m = c(5, 10, 15), design = design, etc.)
  ## ideas:
  ## if user wants to change N and provides data, bootstrap data to that size
}

