% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_functions}
\alias{compare_functions}
\alias{compare_designs}
\alias{compare_design_code}
\alias{compare_design_summaries}
\alias{compare_design_data}
\alias{compare_design_estimates}
\alias{compare_design_estimands}
\title{Compare two designs}
\usage{
compare_designs(design1, design2, format = "ansi8",
  mode = "sidebyside", pager = "off", context = -1L, rmd = FALSE)

compare_design_code(design1, design2, format = "ansi256",
  mode = "sidebyside", pager = "off", context = -1L, rmd = FALSE)

compare_design_summaries(design1, design2, format = "ansi256",
  mode = "sidebyside", pager = "off", context = -1L, rmd = FALSE)

compare_design_data(design1, design2, format = "ansi256",
  mode = "sidebyside", pager = "off", context = -1L, rmd = FALSE)

compare_design_estimates(design1, design2, format = "ansi256",
  mode = "sidebyside", pager = "off", context = -1L, rmd = FALSE)

compare_design_estimands(design1, design2, format = "ansi256",
  mode = "sidebyside", pager = "off", context = -1L, rmd = FALSE)
}
\arguments{
\item{design1}{A design object, typically created using the + operator}

\item{design2}{A design object, typically created using the + operator}

\item{format}{Format (in console or HTML) options from \code{diffobj::diffChr}}

\item{mode}{Mode options from \code{diffobj::diffChr}}

\item{pager}{Pager option from \code{diffobj::diffChr}}

\item{context}{Context option from \code{diffobj::diffChr} which sets the number of lines around differences that are printed. By default, all lines of the two objects are shown. To show only the lines that are different, set \code{context = 0}; to get one line around differences for context, set to 1.}

\item{rmd}{Set to \code{TRUE} use in Rmarkdown HTML output. NB: will not work with LaTeX, Word, or other .Rmd outputs.}
}
\description{
Compare two designs
}
\examples{

design1 <- declare_population(N = 100, u = rnorm(N)) +
  declare_potential_outcomes(Y ~ Z + u) +
  declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(n = 75) +
  declare_assignment(m = 50) +
  declare_reveal(Y, Z) +
  declare_estimator(Y ~ Z, estimand = "ATE")

design2 <- declare_population(N = 200, u = rnorm(N)) +
  declare_potential_outcomes(Y ~ 0.5*Z + u) +
  declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_sampling(n = 100) +
  declare_assignment(m = 25) +
  declare_reveal(Y, Z) +
  declare_estimator(Y ~ Z, model = lm_robust, estimand = "ATE")
 
 compare_designs(design1, design2)
 compare_design_code(design1, design2)
 compare_design_summaries(design1, design2)
 compare_design_data(design1, design2)
 compare_design_estimates(design1, design2)
 compare_design_estimands(design1, design2)

}
