% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_sample_frame.R
\name{covariates_table}
\alias{covariates_table}
\title{Print a table describing covariates described in sample_frame}
\usage{
covariates_table(sample_frame)
}
\arguments{
\item{sample_frame}{a sample_frame object created with declare_sample frame.}
}
\description{
Print a table describing covariates described in sample_frame
}

