% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_population.R
\name{covariates_table}
\alias{covariates_table}
\title{Print a table describing covariates described in population}
\usage{
covariates_table(population)
}
\arguments{
\item{population}{a population object created with declare_population frame.}
}
\description{
Print a table describing covariates described in population
}

