% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_analysis.R
\name{declare_analysis}
\alias{declare_analysis}
\alias{get_estimands}
\alias{get_estimands_model}
\alias{get_estimates}
\alias{get_estimates_model}
\alias{truth_data_frame}
\title{Declare an experimental analysis}
\usage{
declare_analysis(formula, treatment_variable = "Z", outcome_variable = NULL,
  estimator = difference_in_means, subset = NULL, weights = NULL, ...,
  quantity_of_interest = NULL,
  quantity_of_interest_labels = substitute(quantity_of_interest),
  estimand_formula = NULL, estimand = NULL, estimand_subset = NULL,
  estimand_weights = NULL, estimand_options = NULL,
  estimand_quantity_of_interest = NULL,
  estimand_quantity_of_interest_labels = NULL)

truth_data_frame(formula = NULL, treatment_variable = "Z", weights = NULL,
  subset = NULL, data = data, estimand_options = NULL, sep = "_", ...)

get_estimates_model(analysis, data)

get_estimands_model(analysis, data)

get_estimates(analysis, quantity_of_interest = NULL, data,
  analysis_labels = NULL)

get_estimands(analysis, qoi = NULL, data, statistics = "est",
  analysis_labels = NULL)
}
\arguments{
\item{formula}{A standard R formula as a string or formula object, i.e. Y ~ Z, that indicates the outcome and the treatment indicator as well as covariates, block indicators, etc. that are used to construct estimates and/or estimands. By default, the formula provided is used for estimates and estimands.}

\item{treatment_variable}{The variable name of the treatment indicator as a string. Defaults to "Z".}

\item{outcome_variable}{The variable name of the outcome variable as a string. Defaults to "Y".}

\item{subset}{A string indicating the subset of the data to take in estimates and, by default, estimands.}

\item{weights}{A string indicating the name of the weights variable to be used in weighted estimators such as WLS.}

\item{estimand}{defaults to "ATE".}

\item{data}{what is it?}

\item{sep}{what is it?}

\item{analysis}{what is it?}

\item{qoi}{defaults to "ATE".}

\item{statistics}{what is it?}

\item{method}{The method used in the analysis to construct estimates and, by default, estimands. Indicate either string (i.e. "lm" or "glm") or a function object that takes as arguments data, design and spits out a standard R class such as lm or glm. Defaults to "lm".}

\item{formula_estimand}{desc}

\item{method_estimand}{either string (i.e. "lm" or "glm") or a function object that takes as arguments data, design and spits out a standard R class such as lm or glm}

\item{subset_estimand}{A string indicating the subset of the data to take for estimands, if the user wishes to have a different subset used in calculating estimands than the subset used for the estimates.}

\item{weights_estimand}{A string indicating the name of the weights variable to be used in weighted estimators such as WLS for calculating estimands only, if the user desires a different weights variable than used to calculate estimates.}

\item{qoi_only}{desc}

\item{qoi_labels}{desc}

\item{formula}{what is it?}

\item{treatment_variable}{what is it?}

\item{subset}{what is it?}

\item{data}{what is it?}

\item{analysis}{what is it?}

\item{qoi}{what is it?}

\item{data}{what is it?}

\item{analysis}{what is it?}

\item{qoi}{what is it?}

\item{data}{what is it?}
}
\value{
a list containing a function to conduct the analysis and a function to extract the result of the test
}
\description{
Description
}
\examples{
# these examples don't work yet
# declare_analysis(analysis = "diff-in-means")
# declare_analysis(analysis = function(Y, Z, data) lm(paste(Y, "~", Z), data = data))
}

