% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_analysis.R
\name{declare_analysis}
\alias{declare_analysis}
\title{Declare an experimental analysis}
\usage{
declare_analysis(formula, treatment_variable = "Z", method = "lm",
  subset = NULL, weights = NULL, estimand = "ATE",
  formula_estimand = NULL, outcome_variable = NULL, qoi = "ATE",
  qoi_only = FALSE, qoi_labels = NULL, ...)
}
\arguments{
\item{formula}{an optional formula object to define analyses such as linear regressions with covariates}

\item{treatment_variable}{}

\item{method}{either string (i.e. "lm" or "glm") or a function object that takes as arguments data, design and spits out a standard R class such as lm or glm}

\item{subset}{a string indicating which subset of the data to take for analyses}

\item{weights}{}

\item{estimand}{}

\item{formula_estimand}{}

\item{outcome_variable}{}

\item{qoi}{}

\item{qoi_only}{}

\item{...}{additional options to be sent to the analysis function and the test_success function}

\item{method}{}
}
\value{
a list containing a function to conduct the analysis and a function to extract the result of the test
}
\description{
Description
}
\examples{
# these examples don't work yet
# declare_analysis(analysis = "diff-in-means")
# declare_analysis(analysis = function(Y, Z, data) lm(paste(Y, "~", Z), data = data))
}

