% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_assignment.R
\name{declare_assignment}
\alias{declare_assignment}
\alias{assignment_handler}
\alias{assignment_handler_legacy}
\title{Declare Data Strategy: Assignment}
\usage{
declare_assignment(..., handler = assignment_handler_legacy, label = NULL)

assignment_handler(data, ...)

assignment_handler_legacy(
  data,
  ...,
  assignment_variable = "Z",
  append_probabilities_matrix = FALSE
)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}

\item{data}{A data.frame.}

\item{assignment_variable}{Name for assignment variable (quoted). Defaults to "Z". Argument to be used with default handler.}

\item{append_probabilities_matrix}{Should the condition probabilities matrix be appended to the data? Defaults to FALSE.  Argument to be used with default handler.}
}
\value{
A function that takes a data.frame as an argument and returns a data.frame with assignment columns appended.
}
\description{
Declare Data Strategy: Assignment
}
\examples{

# setting up a design stub
design <- declare_model(
  classrooms = add_level(10),
  individuals = add_level(20, female = rbinom(N, 1, 0.5))
) + NULL

# Declare assignment of m units to treatment
design + declare_assignment(Z = complete_ra(N = N, m = 100))

# Declare assignment specifying varying block probabilities
design + 
  declare_assignment(Z = block_ra(blocks = female, 
                                  block_prob = c(1/3, 2/3)))

# Declare assignment of clusters with probability 1/4
design + declare_assignment(
  Z = cluster_ra(prob = 1/4, clusters = classrooms))
 
# Declare factorial assignment (Approach 1): 
#   Use complete random assignment to assign Z1 
#   then block on Z1 to assign Z2. 
design + 
   declare_assignment(Z1 = complete_ra(N = N, m = 100),
                      Z2 = block_ra(blocks = Z1))
   
# Declare factorial assignment (Approach 2): 
#   Assign to four conditions and then split into Z1 and Z2 
design +  
  declare_assignment(Z = complete_ra(N = N, conditions = 1:4),
                     Z1 = as.numeric(Z \%in\% 2:3), 
                     Z2 = as.numeric(Z \%in\% 3:4))
   
# Declare assignment using functions outside randomizr package:
design + 
  declare_assignment(Z = rbinom(n = N, size = 1, prob = 0.35))

}
