% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_assignment.R
\name{declare_assignment}
\alias{declare_assignment}
\title{Declare the experimental assignment}
\usage{
declare_assignment(potential_outcomes, block_variable_name = NULL,
  cluster_variable_name = NULL, m = NULL, m_each = NULL,
  prob_each = NULL, block_m = NULL, block_prob = NULL,
  excluded_arms = NULL, baseline_condition = NULL,
  treatment_variable = "Z", custom_assignment_function = NULL,
  custom_block_function = NULL, custom_cluster_function = NULL,
  existing_assignment_variable_name = NULL)
}
\arguments{
\item{potential_outcomes}{potential_outcomes object, as created by \code{\link{declare_potential_outcomes}} (required).}

\item{m}{the number of units (or clusters) to be assigned to treatment in a two-arm trial.}

\item{m_each}{a vector describing the number of units (or clusters) to be assigned to each treatment arm in a multi-arm trial.  Must sum to N (for individually randomized experments) or N_clusters (for cluster randomized experiments).}

\item{prob_each}{a vector describing the probability of units (or clusters) being assigned to each treatment arm. Must sum to 1.}

\item{block_m}{a matrix with the same number of rows as blocks and the same number of columns as treatment arms. Cell entries are the number of units (or clusters) to be assigned to each treatment arm.}

\item{block_prob}{a matrix with the same number of rows as blocks and the same number of columns as treatment arms. Cell entries are the probabilities of assignment to each treatment arm.}

\item{excluded_arms}{a character vector excluding some potential outcomes from the randomization.  Used primarily when comparing assignments that feature different numbers of treatment arms.}

\item{blocks}{blocks object, as created by \code{\link{declare_blocks}} (optional).}

\item{clusters}{clusters object, as created by \code{\link{declare_clusters}} (optional).}
}
\value{
assignment object
}
\description{
Declare the experimental assignment
}

