% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_citation.R
\name{declare_citation}
\alias{declare_citation}
\alias{citation_handler}
\title{Declare research study citation}
\usage{
declare_citation(..., handler = citation_handler, label = NULL)

citation_handler(data = stop("Don't evaluate data argument"), title = NULL,
  authors = NULL, description = "Unpublished research design declaration.",
  timestamp = Sys.time(), citation = NULL)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}

\item{data}{Do not use}

\item{title}{(optional) The title of the study, as a character string.}

\item{authors}{(optional) The authors of the study, as a character string.}

\item{description}{(optional) A description of the design in words, as a character string, stored alongside the declaration in code.}

\item{timestamp}{a timestamp to use in the citation}

\item{citation}{(optional) The preferred citation for the design, as a character string. Either include the full citation in text, or paste a BibTeX entry. If title and authors are specified and you leave citation empty, a BibTeX entry will be created automatically.}
}
\value{
a function that takes a data.frame as an argument and returns citation data
}
\description{
Declare research study citation
}
\examples{

 cite <- declare_citation(
    title = paste("Reducing intergroup prejudice and conflict using the media:", 
    "A field experiment in Rwanda"), 
    author = "Paluck, E.L.", 
    description = "Journal of Personality and Social Psychology, 96, 574-587.")
 design <- cite + NULL
 cite_design(design)
}
