% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_citation.R
\name{declare_citation}
\alias{declare_citation}
\title{Declare Research Study Citation}
\usage{
declare_citation(..., handler = citation_function_default, label = NULL)
}
\arguments{
\item{...}{Arguments to the citation function}

\item{handler}{A function that takes a data.frame and yield citation data}
}
\value{
a function that takes a data.frame as an argument and returns citation data
}
\description{
Declare Research Study Citation
}
\details{

}
