% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_design.R
\name{declare_design}
\alias{declare_design}
\title{Declare the design}
\usage{
declare_design(population, sampling = NULL, assignment, estimator = NULL,
  potential_outcomes, inputs = NULL, label = NULL,
  text_description = NULL)
}
\arguments{
\item{population}{A population object, as created by \code{\link{declare_population}}.}

\item{sampling}{A sampling object, as created by \code{\link{declare_sampling}}.}

\item{assignment}{An assignment object, as created by \code{\link{declare_assignment}}.}

\item{estimator}{An estimator object, as created by \code{\link{declare_estimator}}. Optionally a list of estimator objects.}

\item{potential_outcomes}{A potential_outcomes object, as created by \code{\link{declare_potential_outcomes}}. Optionally a list of potential_outcomes objects.}

\item{label}{An optional design label.}
}
\value{
A design object
}
\description{
This function gathers together the declarations for population, sampling, assignment, estimator(s), and potential outcomes. Design objects can be summarized with summary(my_design), diagnosed with diagnose_design(my_design), and modified with modify_design(my_design).
}

