% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_design.R
\name{declare_design}
\alias{declare_design}
\title{Declare the experimental design}
\usage{
declare_design(potential_outcomes, blocks = NULL, clusters = NULL,
  m = NULL, m_each = NULL, prob_each = NULL, block_m = NULL,
  excluded_arms = NULL)
}
\arguments{
\item{potential_outcomes}{potential_outcomes object, as created by \code{\link{declare_potential_outcomes}} (required).}

\item{blocks}{blocks object, as created by \code{\link{declare_blocks}} (optional).}

\item{clusters}{clusters object, as created by \code{\link{declare_clusters}} (optional).}

\item{m}{the number of units (or clusters) to be assigned to treatment in a two-arm trial.}

\item{m_each}{a vector describing the number of units (or clusters) to be assigned to each treatment arm in a multi-arm trial.  Must sum to N (for individually randomized experments) or N_clusters (for cluster randomized experiments).}

\item{prob_each}{a vector describing the probability of units (or clusters) being assigned to each treatment arm. Must sum to 1.}

\item{block_m}{a matrix with the same number of rows as blocks and the same number of columns as treatment arms. Cell entries are the number of units (or clusters) to be assigned to each treatment arm.}

\item{excluded_arms}{a character vector excluding some potential outcomes from the randomization.  Used primarily when comparing designs that feature different numbers of treatment arms.}
}
\value{
design object
}
\description{
Declare the experimental design
}

