% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_diagnosand.R
\name{declare_diagnosand}
\alias{declare_diagnosand}
\title{Diagnosand}
\usage{
declare_diagnosand(diagnostic_statistic_text, diagnostic_statistic_function,
  summary_function = mean, label = NULL, description = NULL,
  na_function = na.pass, ...)
}
\arguments{
\item{diagnostic_statistic_text}{}

\item{diagnostic_statistic_function}{}

\item{summary_function}{}

\item{label}{}

\item{description}{}

\item{...}{}
}
\description{
Diagnosand
}

