% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_diagnosands.R
\name{declare_diagnosands}
\alias{declare_diagnosands}
\title{Declare Diagnosands}
\usage{
declare_diagnosands(...)
}
\arguments{
\item{...}{A set of diagnosands, i.e. bias = mean(est - estimand). The diagnostic statistic (i.e., est - estimand) and its associated summary function (i.e., mean) are declared at the same time.}
}
\value{
a function that returns a data.frame
}
\description{
Declare Diagnosands
}
\details{
Diagnosands summarize the simulations generated by \code{\link{diagnose_design}}.  Typically, the columns of the resulting simulations data.frame include the following variables: est, se, p, ci_lower, ci_upper, and estimand. Most diagnosands will be some function of these variables.

If you do not specify a particular set of diagnosands, the following diagnosands will be reported by default:
bias = mean(est - estimand)
rmse = sqrt(mean((est - estimand)^2))
power = mean(p < .05)
coverage = mean(estimand <= ci_upper & estimand >= ci_lower)
mean_estimate = mean(est)
sd_estimate = sd(est)
type_s_rate = mean((sign(est) != sign(estimand)) & p < .05)
mean_estimand = mean(estimand)
}
\examples{

default_diagnosands <- declare_diagnosands(
 bias = mean(est - estimand),
 rmse = sqrt(mean((est - estimand)^2)),
 power = mean(p < .05),
 coverage = mean(estimand <= ci_upper & estimand >= ci_lower),
 mean_estimate = mean(est),
 sd_estimate = sd(est),
 type_s_rate = mean((sign(est) != sign(estimand)) & p < .05),
 mean_estimand = mean(estimand))

}
