% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_diagnosands.R
\name{diagnosand_handler}
\alias{diagnosand_handler}
\alias{declare_diagnosands}
\title{Declare Diagnosands}
\usage{
diagnosand_handler(data, ..., select, subtract, keep_defaults = TRUE,
  subset = NULL, alpha = 0.05, label)

declare_diagnosands(..., handler = diagnosand_handler, label = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{...}{A set of new diagnosands}

\item{select}{A set of the default diagnosands to report e.g., \code{select = c(bias, rmse)}}

\item{subtract}{A set of the default diagnosands to NOT report e.g., \code{subtract = c(bias, rmse)}}

\item{keep_defaults}{A flag for whether to report the default diagnosands. Defaults to \code{TRUE}.}

\item{subset}{A subset of the simulations data frame within which to calculate diagnosands e.g. \code{subset = p < .05}.}

\item{alpha}{Alpha significance level. Defaults to \code{.05}.}

\item{label}{Label for the set of diagnosands.}

\item{handler}{a tidy-in, tidy-out function}
}
\value{
a function that returns a data.frame
}
\description{
Declare Diagnosands
}
\details{
If coefficients is TRUE, the names of ... will be returned in a `coefficients` column, and `estimand_label`
will contain the step label. This can be used as an additional dimension for use in diagnosis.

Diagnosands summarize the simulations generated by \code{\link{diagnose_design}}.  Typically, the columns of the resulting simulations data.frame include the following variables: est, se, p, ci_lower, ci_upper, and estimand. Most diagnosands will be some function of these variables.

By default (\code{keep_defaults = TRUE}), a set of common diagnosands are reported:

bias = mean(est - estimand)\cr
rmse = sqrt(mean((est - estimand)^2))\cr
power = mean(p < .05)\cr
coverage = mean(estimand <= ci_upper & estimand >= ci_lower)\cr
mean_estimate = mean(est)\cr
sd_estimate = sd(est)\cr
type_s_rate = mean((sign(est) != sign(estimand))[p < alpha])\cr
mean_estimand = mean(estimand)\cr
}
\examples{

my_population <- declare_population(N = 500, noise = rnorm(N))

my_potential_outcomes <- declare_potential_outcomes(
  Y_Z_0 = noise, Y_Z_1 = noise +
  rnorm(N, mean = 2, sd = 2))

my_assignment <- declare_assignment()

my_estimand <- declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0))

my_estimator <- declare_estimator(Y ~ Z, estimand = my_estimand)

my_reveal <- declare_reveal()

design <- my_population + my_potential_outcomes + my_estimand +
        my_assignment + my_reveal + my_estimator

\dontrun{
# using built-in defaults:
diagnosis <- diagnose_design(design)
diagnosis
}

# You can select a set of those diagnosands via the \\code{select} argument e.g., 

my_diagnosands <- declare_diagnosands(select = c(bias, rmse))

# Alternatively, you can report all of the default diagnosands and subtract a subset of them e.g.,

my_diagnosands <- declare_diagnosands(subtract = type_s_rate)

# You can add your own diagnosands in addition to or instead of the defaults e.g.,

my_diagnosands <- declare_diagnosands(median_bias = median(est - estimand))

# or to report only \\code{median_bias}

my_diagnosands <- declare_diagnosands(median_bias = median(est - estimand), keep_defaults = FALSE)

# Below is the code that makes the default diagnsoands.
# You can use these as a model when writing your own diagnosands.

default_diagnosands <- declare_diagnosands(
 bias = mean(est - estimand),
 rmse = sqrt(mean((est - estimand)^2)),
 power = mean(p < .05),
 coverage = mean(estimand <= ci_upper & estimand >= ci_lower),
 mean_estimate = mean(est),
 sd_estimate = sd(est),
 type_s_rate = mean((sign(est) != sign(estimand)) & p < .05),
 mean_estimand = mean(estimand))

}
