% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_estimand_diagnosands.R
\name{declare_diagnosands}
\alias{declare_diagnosands}
\title{Declare Diagnosands}
\usage{
declare_diagnosands(..., handler = diagnosand_handler, label = NULL)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}
}
\value{
a function that returns a data.frame
}
\description{
Declare Diagnosands
}
\details{
Diagnosands summarize the simulations generated by \code{\link{diagnose_design}}.  Typically, the columns of the resulting simulations data.frame include the following variables: est, se, p, ci_lower, ci_upper, and estimand. Most diagnosands will be some function of these variables.

If you do not specify a particular set of diagnosands, the following diagnosands will be reported by default:

bias = mean(est - estimand)\cr
rmse = sqrt(mean((est - estimand)^2))\cr
power = mean(p < .05)\cr
coverage = mean(estimand <= ci_upper & estimand >= ci_lower)\cr
mean_estimate = mean(est)\cr
sd_estimate = sd(est)\cr
type_s_rate = mean((sign(est) != sign(estimand))[p < alpha])\cr
mean_estimand = mean(estimand)\cr
}
\examples{

my_population <- declare_population(N = 500, noise = rnorm(N))

my_potential_outcomes <- declare_potential_outcomes(
  Y_Z_0 = noise, Y_Z_1 = noise +
  rnorm(N, mean = 2, sd = 2))

my_assignment <- declare_assignment()

my_estimand <- declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0))

my_estimator <- declare_estimator(Y ~ Z, estimand = my_estimand)

my_reveal <- declare_reveal()

design <- declare_design(my_population,
                         my_potential_outcomes,
                         my_estimand,
                         my_assignment,
                         my_reveal,
                         my_estimator)

\dontrun{
# using built-in defaults:
diagnosis <- diagnose_design(design)
diagnosis
}

# using a user-defined diagnosand
my_diagnosand <- declare_diagnosands(absolute_error = mean(abs(est - estimand)))

\dontrun{
diagnosis <- diagnose_design(design, diagnosands = my_diagnosand)
diagnosis
}

# this is the code that makes the default diagnsoands.
# Use these as a model when writing your own diagnosands.

default_diagnosands <- declare_diagnosands(
 bias = mean(est - estimand),
 rmse = sqrt(mean((est - estimand)^2)),
 power = mean(p < .05),
 coverage = mean(estimand <= ci_upper & estimand >= ci_lower),
 mean_estimate = mean(est),
 sd_estimate = sd(est),
 type_s_rate = mean((sign(est) != sign(estimand)) & p < .05),
 mean_estimand = mean(estimand))

}
