% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_estimand.R
\name{declare_estimand}
\alias{declare_estimand}
\alias{declare_estimands}
\alias{estimand_handler}
\title{Declare estimand}
\usage{
declare_estimand(..., handler = estimand_handler, label = "estimand")

declare_estimands(..., handler = estimand_handler, label = "estimand")

estimand_handler(data, ..., subset = NULL, term = FALSE, label)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}

\item{data}{a data.frame}

\item{subset}{a subset expression}

\item{term}{TRUE/FALSE}
}
\value{
an estimand declaration, which is a function that accepts a data.frame as an argument and returns a data.frame containing the value of the estimand.
}
\description{
Declares estimands. Estimands are the subjects of inquiry and can be estimated by an estimator.
}
\details{
For the default diagnosands, the return value of the handler should have \code{estimand_label} and \code{estimand} columns.

If term is TRUE, the names of ... will be returned in a \code{term} column,
and \code{estimand_label} will contain the step label. This can be used as
an additional dimension for use in diagnosis.
}
\examples{

# Set up a design for use in examples:

design <-
  declare_population(N = 100, X = rnorm(N)) +
  declare_potential_outcomes(Y ~ (.25 + X) * Z + rnorm(N)) +
  declare_assignment(m = 50) 

design + declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0))


design + declare_estimand(ATT = mean(Y_Z_1 - Y_Z_0),
                                   subset = (Z == 1))

# Add estimands to a design along with estimators that reference them

design_1 <-
  design +
  declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_estimator(Y ~ Z, estimand = "ATE")

run_design(design_1)

# Two estimands, one estimator

design_2 <-
  design +
  declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0)) +
  declare_estimand(ATT = mean(Y_Z_1 - Y_Z_0), subset = (Z == 1)) + 
  declare_estimator(Y ~ Z, estimand = c("ATE", "ATT"))

run_design(design_2)

# Two estimands, two coefficients from one estimator

design_3 <-
  design +
  declare_estimand(intercept = mean(Y_Z_0),
                   slope = mean(Y_Z_1 - Y_Z_0)) +
  declare_estimator(
    Y ~ Z,
    model = lm_robust,
    term = TRUE,
    estimand = c("intercept", "slope")
  )

run_design(design_3)

}
