% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_estimand.R
\name{declare_estimand}
\alias{declare_estimand}
\title{Declare Estimand}
\usage{
declare_estimand(..., handler = estimand_function_default,
  label = "my_estimand")
}
\arguments{
\item{...}{Arguments to the estimand function. For example, you might specify ATE = mean(Y_Z_1 - Y_Z_0), which would declare the estimand to be named ATE and to be the mean of the difference in the control and treatment potential outcome.}

\item{handler}{A function that takes a data.frame as an argument and returns a data.frame with the estimand and a label. By default, the estimand function accepts an expression such as ATE = mean(Y_Z_1-Y_Z_0).}

\item{label}{An optional label to name the estimand, such as ATE. Typically, the label is inferred from how you specify the estimand in \code{...}, i.e. if you specify ATE = mean(Y_Z_1 - Y_Z_0) the estimand label will be ATE.}
}
\value{
a function that accepts a data.frame as an argument and returns a data.frame containing the value of the estimand.
}
\description{
Declare Estimand
}
\examples{

my_population <- declare_population(N = 100)

my_potential_outcomes <- declare_potential_outcomes(
  formula = Y ~ .25 * Z,
  condition_names = c(0, 1))

df <- my_potential_outcomes(my_population())

# Use the default estimand setup to
# declare an average treatment effect estimand

my_estimand <- declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0))

my_estimand(df)

# Custom random assignment functions

my_estimand_function <- function(data, label) {
  ret <- with(data, median(Y_Z_1 - Y_Z_0))
  data.frame(setNames(ret, label))
}
my_estimand_custom <- declare_estimand(
  handler = my_estimand_function, label = "medianTE")

my_estimand_custom(df)

}
