% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_estimator.R
\name{declare_estimator}
\alias{declare_estimator}
\title{Estimator declaration}
\usage{
declare_estimator(formula = NULL, model = NULL, model_options = NULL,
  estimates, estimates_options = NULL, subset = NULL,
  weights_variable_name = NULL, labels = NULL, text_description = NULL,
  estimand = NULL)
}
\arguments{
\item{formula}{An object of class "formula".}

\item{model}{}

\item{model_options}{}

\item{estimates}{}

\item{estimates_options}{}

\item{subset}{}

\item{weights_variable_name}{}

\item{labels}{}

\item{estimand}{}
}
\description{
Estimator declaration
}

