% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{declare_internal_inherit_params}
\alias{declare_internal_inherit_params}
\title{Used to inherit roxygen docs}
\usage{
declare_internal_inherit_params(
  ...,
  handler = function(data, ...) data.frame(BLNK = "MSG", stringsAsFactors = TRUE),
  label = NULL
)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}
}
\description{
Used to inherit roxygen docs
}
\keyword{internal}
