% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_model.R
\name{declare_model}
\alias{declare_model}
\title{Declare the size and features of the population}
\usage{
declare_model(..., handler = fabricate, label = NULL)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}
}
\value{
A function that returns a data.frame.
}
\description{
Declare the size and features of the population
}
\examples{

# Default handler is fabricatr::fabricate

# Declare a single-level population with no covariates
my_population <- declare_model(N = 100)

# Declare a population from existing data
my_population <- declare_model(sleep)

# Declare a single-level population with a covariate
my_population <- declare_model(
  N = 6,
  gender = draw_binary(N, prob = 0.5),
  height_ft = rnorm(N, mean = 5 + 4/12 + 4/12 * gender, sd = 3/12)
)
my_population()

# Declare a two-level hierarchical population
# containing cities within regions and a
# pollution variable defined at the city level

my_population <- declare_model(
  regions = add_level(N = 5),
  cities = add_level(N = 10, pollution = rnorm(N, mean = 5))
)
my_population()

# Declare a population using a custom function

my_population_function <- function(N) {
  data.frame(u = rnorm(N))
}

my_population_custom <- declare_model(N = 10, handler = my_population_function)

my_population_custom()
}
