% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_noncompliance.R
\name{declare_noncompliance}
\alias{declare_noncompliance}
\title{Define the potential outcomes of the experiment}
\usage{
declare_noncompliance(condition_names = NULL, outcome_formula = NULL,
  population_proportions = NULL, proportion_outcome_name = NULL)
}
\arguments{
\item{condition_names}{A character vector indicating the names of the conditions to which subjects can be assigned. Conceptually identical to the potential outcomes that are revealed in each condition, implicitly invoking SUTVA.}

\item{outcome_formula}{A regression-like expression for declaring the relationship between the outcome, treatments, and optionally, the covariates.}

\item{population_proportions}{what is it?}

\item{proportion_outcome_name}{what is it?}
}
\value{
outcomes_object
}
\description{
Define the potential outcomes of the experiment
}

