% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_population.R
\name{declare_population}
\alias{declare_population}
\title{Declare the size and features of the population}
\usage{
declare_population(..., handler = fabricate, label = NULL)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}
}
\value{
A potential outcomes declaration, which is a function that returns a data.frame.
}
\description{
Deprecated. Please use declare_model instead.
}
\keyword{internal}
