% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_population.R
\name{declare_population}
\alias{declare_population}
\title{Declare the Size and Features of the Population}
\usage{
declare_population(..., handler = fabricate, label = NULL)
}
\arguments{
\item{...}{Arguments to the population function}

\item{handler}{A function that generates a data.frame containing features of the population. By default, the \link{fabricatr} function \code{\link{fabricate}} function is used, which allows you to draw single or multi-level hierarchical population data.}
}
\value{
a function that returns a data.frame
}
\description{
Declare the Size and Features of the Population
}
\examples{

# Declare a single-level population with no covariates
my_population <- declare_population(N = 100)
head(my_population())

# Declare a single-level population with a covariate
my_population <- declare_population(
  N = 100,
  height_ft = runif(N, 3.5, 8)
)
head(my_population())

# Declare a two-level hierarchical population
# containing cities within regions

pop <- declare_population(
 regions = level(N = 5),
 cities = level(N = 10, pollution = rnorm(N, mean = 5)))
head(my_population())

# Custom population functions

my_population_function <- function(N) {
  data.frame(u = rnorm(N))
}

my_population_custom <- declare_population(
  handler = my_population_function, N = 100)

head(my_population_custom())
}
