% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_population.R
\name{declare_population}
\alias{declare_population}
\title{New Declare the Study Population}
\usage{
declare_population(..., size, global_transformations = NULL,
  other_arguments = NULL, level_IDs = NULL, super_population = FALSE,
  random_seed = 42, data = NULL, custom_population_function = NULL,
  make_unique_ID = FALSE)
}
\arguments{
\item{...}{A named list of expressions or list of named lists of expressions that generate data. These should generate vectors, the privileged name n_ can be used to specify the leve-specific N (i.e. rnorm(n_) creates 100 normal draws if it is at the level of 100 individuals, or 10 normal draws if it is at the level of 10 cities). These must be declared consecutively and are level-specific (see global_transformations for expressions that can be evaluated over the entire dataset).}

\item{size}{A scalar (for single-level datasets) indicating the N, a vector of N's for each level (i.e. c(100,50,2) for 100 individuals in 50 cities in 2 regions), or a list of vectors specifying how many of the lower-level units are in each unit at that level (i.e. list(rep(1,100),rep(2,50),rep(25,2)) for the same example).}

\item{global_transformations}{An optional named list of expressions that can be evaluated across levels. For example, if group-invariant means and SDs have been created, these can be used with an expression in global_transformations in order to create individual-level draws of these group-level parameters.}

\item{other_arguments}{An optional list of objects that are used by the expressions passed to ... and global_transformations.}

\item{level_IDs}{An optional list of level ID indicators that are otherwise inferred from ... or generated using a default.}

\item{super_population}{If TRUE, data is thought of as a single draw from a super-population, and data-resampling is performed during the diagnostics phase. If FALSE, the population is thought of as finite.}

\item{random_seed}{The random seed used for re-sampling.}

\item{data}{User-provided data for bootrstrapping.}

\item{custom_population_function}{User-provided function for regenerating data.}

\item{make_unique_ID}{If TRUE, an ID is made for each unit that indicates all of the other level IDs.}
}
\description{
New Declare the Study Population
}
\examples{

# Lots of nice examples go here.

}

