% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_potential_outcomes.R
\name{declare_potential_outcomes}
\alias{declare_potential_outcomes}
\title{Declare potential outcomes}
\usage{
declare_potential_outcomes(
  ...,
  handler = potential_outcomes_handler,
  label = NULL
)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}
}
\value{
a function that returns a data.frame
}
\description{
Deprecated. Please use the potential_outcomes function within a declare_model declaration.
}
\keyword{internal}
