% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_potential_outcomes.R
\name{declare_potential_outcomes}
\alias{declare_potential_outcomes}
\title{Potential Outcomes}
\usage{
declare_potential_outcomes(..., handler = potential_outcomes_function_default,
  label = NULL)
}
\arguments{
\item{...}{Arguments to the potential_outcomes_function}

\item{potential_outcomes_function}{A function that accepts a data.frame as an argument and returns a data.frame with potential outcomes columns appended. See the examples for the behavior of the default function.}
}
\value{
a function that returns a data.frame
}
\description{
Potential Outcomes
}
\details{
A `declare_potential_outcomes` declaration returns a function. That function takes data and returns data with potential outcomes columns appended. These columns describe the outcomes that each unit would express if that unit were in the corresponding treatment condition.

The potential outcomes function can sometimes be a stumbling block for users, as some are uncomfortable asserting anything in particular about the very causal process that they are conducting a study to learn about! We recommend trying to imagine what your preferred theory would predict, what an alternative theory would predict, and what your study would reveal if there were no differences in potential outcomes for any unit (i.e., all treatment effects are zero).
}
\examples{
my_population <-
declare_population(N = 1000,
                   income = rnorm(N),
                   age = sample(18:95, N, replace = TRUE))
pop <- my_population()

# By default, there are two ways of declaring potential outcomes:
# as separate variables or using a formula:

# As separate variables

my_potential_outcomes <-
     declare_potential_outcomes(
        Y_Z_0 = .05,
        Y_Z_1 = .30 + .01 * age)

head(my_potential_outcomes(pop))

# Using a formula
 my_potential_outcomes <- declare_potential_outcomes(
 formula = Y ~ .25 * Z + .01 * age * Z)
 pop_pos <- my_potential_outcomes(pop)
 head(pop_pos)

 # condition_names defines the "range" of the potential outcomes function
 my_potential_outcomes <-
      declare_potential_outcomes(
      formula = Y ~ .25 * Z + .01 * age * Z,
      condition_names = 1:4)

head(my_potential_outcomes(pop))

}
