% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_potential_outcomes.R
\name{declare_potential_outcomes}
\alias{declare_potential_outcomes}
\title{Define the potential outcomes of the experiment}
\usage{
declare_potential_outcomes(condition_names, outcome_formula,
  outcome_variable_DGP = declare_variable(), cluster_variable = NULL,
  ICC = NULL)
}
\arguments{
\item{covariate_object}{covariates data frame}
}
\value{
outcomes_object
}
\description{
Define the potential outcomes of the experiment
}

