% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_potential_outcomes.R
\name{declare_potential_outcomes}
\alias{declare_potential_outcomes}
\title{Define the potential outcomes of the experiment}
\usage{
declare_potential_outcomes(condition_names = NULL, outcome_formula = NULL,
  population_proportions = NULL, proportion_outcome_name = NULL,
  sep = "_")
}
\arguments{
\item{condition_names}{A character vector indicating the names of the conditions to which subjects can be assigned. Conceptually identical to the potential outcomes that are revealed in each condition, implicitly invoking SUTVA.}

\item{outcome_formula}{A regression-like expression for declaring the relationship between the outcome, treatments, and optionally, the covariates.}

\item{population_proportions}{For binary outcomes, a vector of proportions of successes is sufficient. For multinomial outcomes, provide a matrix of proportions, whose rows indicate the possible outcomes, and whose columns sum to one and indicate the treatment conditions. Outcome and condition names are automatically inherited from the rows and columns of this matrix, where possible. Note: if the user decides to define the treatment effects in terms of population proportions, individual-level data-generating processes, such as those described by the outcome_formula, will no longer be possible.}

\item{proportion_outcome_name}{The name of the outcome variable when using population proportions to define the outcomes.}
}
\value{
outcomes_object
}
\description{
Define the potential outcomes of the experiment
}

