% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_potential_outcomes.R
\name{declare_potential_outcomes}
\alias{declare_potential_outcomes}
\title{Create the potential outcomes of the experiment}
\usage{
declare_potential_outcomes(covariate_object, design_object, outcome_formula,
  outcome_DGP = declare_DGP(), cluster_variable = NULL,
  N_per_cluster = NULL, ICC = 0.01, unit_variance = 1)
}
\arguments{
\item{covariate_object}{covariates data frame}
}
\value{
a data frame?
}
\description{
Create the potential outcomes of the experiment
}

