% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_potential_outcomes.R
\name{declare_potential_outcomes}
\alias{declare_potential_outcomes}
\title{Declare potential outcomes}
\usage{

  declare_potential_outcomes(potential_outcomes_function = default_potential_outcomes_function,
  formula = NULL, outcome_variable_name = NULL, condition_names = NULL,
  sep = "_", assignment_variable_name = "Z", interference = NULL,
  attrition = NULL, ..., description = NULL)
}
\arguments{
\item{potential_outcomes_function}{}

\item{formula}{}

\item{outcome_variable_name}{}

\item{condition_names}{}

\item{sep}{}

\item{assignment_variable_name}{}

\item{options}{}
}
\value{
potential_outcomes object.
}
\description{
Declare potential outcomes
}

