% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_sampling.R
\name{declare_sampling}
\alias{declare_sampling}
\alias{sampling_handler}
\alias{sampling_handler_legacy}
\title{Declare sampling procedure}
\usage{
declare_sampling(..., handler = sampling_handler, label = NULL)

sampling_handler(data, ..., filter = S == 1)

sampling_handler_legacy(
  data,
  ...,
  sampling_variable = "S",
  drop_nonsampled = TRUE
)
}
\arguments{
\item{...}{arguments to be captured, and later passed to the handler}

\item{handler}{a tidy-in, tidy-out function}

\item{label}{a string describing the step}

\item{data}{A data.frame.}

\item{filter}{Unquoted expression for filtering S. By default subsets to \code{S == 1}.}

\item{sampling_variable}{The prefix for the sampling inclusion probability variable.}

\item{drop_nonsampled}{Logical indicating whether to drop units that are not sampled. Default is \code{TRUE}.}
}
\value{
A function that takes a data.frame as an argument and returns a data.frame subsetted to sampled observations and (optionally) augmented with inclusion probabilities and other quantities.
}
\description{
Declare sampling procedure
}
\examples{

design <- declare_model(

  classrooms = add_level(10),
  individuals = add_level(20, female = rbinom(N, 1, 0.5))

) + NULL

# Complete random sampling

design + declare_sampling(S = complete_rs(N = N, n = 50), filter = S == 1)

# equivalently, by default filter is set to S == 1
design + declare_sampling(S = complete_rs(N = N, n = 50))

# Stratified random sampling

design + declare_sampling(S = strata_rs(strata = female))
}
