% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_sampling.R
\name{declare_sampling}
\alias{declare_sampling}
\title{Declare Sampling Procedure}
\usage{
declare_sampling(..., sampling_function = sampling_function_default)
}
\arguments{
\item{...}{Arguments to the sampling function}

\item{sampling_function}{A function that takes a data.frame, subsets to sampled observations and optionally adds sampling probabilities or other relevant quantities, and returns a data.frame. By default, the sampling_function uses the \code{randomizr} functions \code{\link{draw_rs}} and \code{\link{obtain_inclusion_probabilities}} to conduct random sampling and obtain the probability of inclusion in the sample.}
}
\value{
a function that takes a data.frame as an argument and returns a data.frame subsetted to sampled observations and (optionally) augmented with inclusion probabilities and other quantities.
}
\description{
Declare Sampling Procedure
}
\examples{

my_population <- declare_population(N = 100)
df <- my_population()

# Simple random sampling using randomizr
# use any arguments you would use in draw_rs.

my_sampling <- declare_sampling(n = 50)
df <- my_sampling(df)
dim(df)
head(df)

# Custom random assignment functions

df <- my_population()

my_sampling_function <- function(data) {
   data$S <- rbinom(n = nrow(data),
     size = 1,
     prob = 0.5)
   data[data$S == 1, ]
   }

my_sampling_custom <- declare_sampling(
   sampling_function = my_sampling_function)

df <- my_sampling_custom(df)
dim(df)
head(df)
}
