% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_sampling.R
\name{declare_sampling}
\alias{declare_sampling}
\title{Declare the sampling strategy}
\usage{
declare_sampling(probability = NULL, strata_variable_name = NULL,
  cluster_variable_name = NULL, n = NULL, strata_n = NULL,
  strata_probabilities = NULL, custom_sampling_function = NULL,
  custom_stratification_function = NULL, custom_clustering_function = NULL,
  description = NULL)
}
\arguments{
\item{probability}{The probability of inclusion in the sample. Must be a scalar that takes values between 0 and 1.}

\item{strata_variable_name}{The name of the variable according to which stratified random sampling should be conducted.}

\item{cluster_variable_name}{The name of the variable according to which clustered random sampling should be conducted.}

\item{n}{The number of units (or clusters) to be sampled.}

\item{strata_n}{A vector of the number of units in each stratum to be sampled.}

\item{strata_probabilities}{}

\item{custom_sampling_function}{}

\item{custom_stratification_function}{}

\item{custom_clustering_function}{}
}
\description{
Declare the sampling strategy
}

