% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_sampling.R
\name{declare_sampling}
\alias{declare_sampling}
\title{Declare the sampling strategy}
\usage{
declare_sampling(prob = NULL, strata_variable_name = NULL,
  cluster_variable_name = NULL, n = NULL, strata_n = NULL,
  strata_prob = NULL, custom_sampling_function = NULL,
  custom_strata_function = NULL, custom_cluster_function = NULL)
}
\arguments{
\item{prob}{The probability of inclusion in the sample. Must be a scalar that takes values between 0 and 1.}

\item{strata_variable_name}{The name of the variable according to which stratified random sampling should be conducted.}

\item{cluster_variable_name}{The name of the variable according to which clustered random sampling should be conducted.}

\item{n}{The number of units (or clusters) to be sampled.}

\item{strata_n}{A matrix with the same number of rows as blocks and the same number of columns as treatment arms. Cell entries are the number of units (or clusters) to be assigned to each treatment arm.}

\item{strata_prob}{}

\item{custom_sampling_function}{}

\item{custom_strata_function}{}

\item{custom_cluster_function}{}
}
\description{
Declare the sampling strategy
}

