% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_step.R
\name{declare_step}
\alias{declare_step}
\title{Declare a custom step}
\usage{
declare_step(..., handler = fabricate, label = NULL)
}
\value{
a function that returns a data.frame
}
\description{
Declare a custom step
}
