% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_variable.R
\name{declare_variable}
\alias{declare_variable}
\title{Declare the data-generating process of a variable}
\usage{
declare_variable(linear_mean = NULL, linear_sd = NULL,
  binary_probability = NULL, binary_categories = NULL,
  multinomial_probabilities = NULL, multinomial_categories = NULL)
}
\arguments{
\item{linear_mean}{If the variable is linear}
}
\description{
Declare the data-generating process of a variable
}

