% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_variable.R
\name{declare_variable}
\alias{declare_variable}
\title{Declare the data-generating process of a variable}
\usage{
declare_variable(normal_mean = NULL, normal_sd = NULL,
  binary_probability = NULL, binary_categories = NULL,
  multinomial_probabilities = NULL, multinomial_categories = NULL,
  transformation = NULL)
}
\arguments{
\item{normal_mean}{If the variable is normally distributed, its mean}

\item{normal_sd}{If the variable is normally distributed, its standard deviation}

\item{binary_probability}{How does this work?}

\item{binary_categories}{How does this work?}

\item{multinomial_probabilities}{How does this work?}

\item{multinomial_categories}{How does this work?}

\item{transformation}{How does this work?}
}
\description{
Declare the data-generating process of a variable
}

