% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_variable.R
\name{declare_variable}
\alias{declare_variable}
\title{Declare the data-generating process of a variable}
\usage{
declare_variable(normal_mean = NULL, normal_sd = NULL,
  binary_probability = NULL, binary_categories = NULL,
  multinomial_probabilities = NULL, multinomial_categories = NULL,
  transformation = NULL)
}
\arguments{
\item{normal_mean}{If the variable is normally distributed, its mean}

\item{normal_sd}{If the variable is normally distributed, its standard deviation}

\item{binary_probability}{The probability of success when the outcome is binary.}

\item{binary_categories}{The failure and success labels, respectively.}

\item{multinomial_probabilities}{A vector of probabilities of different draws, as long as there are categories.}

\item{multinomial_categories}{A vector of labels for the multinomial draws}

\item{transformation}{An optional string argument that can perform any operation on the covariates in \code{\link{declare_population}}, provided that: the transformations return a vector of the same length as the dataframe; the transformations only involve variables on one level; and the transformations take place successively (i.e. only variables or transformations which are declared in previous lines can themselves be transformed).}
}
\description{
Declare the data-generating process of a variable
}

