% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_diagnosands.R
\name{default_diagnosands}
\alias{default_diagnosands}
\title{Default diagnosands}
\usage{
default_diagnosands()
}
\value{
A diagnosand \code{design_step}.
}
\description{
Returns the standard set of diagnosands: mean estimand, mean estimate,
bias, SD of estimates, RMSE, power, and coverage of nominal 95 percent
confidence intervals.
}
\examples{
default_diagnosands()
}
