% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{delete_step}
\alias{delete_step}
\title{Delete a step from a design}
\usage{
delete_step(design, step)
}
\arguments{
\item{design}{A \code{design}.}

\item{step}{A label, integer, or \code{design_step}.}
}
\value{
A \code{design}.
}
\description{
Delete a step from a design
}
\examples{
design <- declare_model(N = 30, Y = rnorm(N)) +
  declare_inquiry(mu = mean(Y))
delete_step(design, "inquiry") |> names()
}
