% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs_methods.R
\name{design_diff}
\alias{design_diff}
\title{design_diff
Visualize differences between design_comparison objects or summaries with library(diffobj). Appears in Viewer tab of RStudio.}
\usage{
design_diff(reference, alternative)
}
\arguments{
\item{reference}{The reference object (a.k.a., 'target').}

\item{alternative}{Another object to compare the referencd to.}
}
\description{
design_diff
Visualize differences between design_comparison objects or summaries with library(diffobj). Appears in Viewer tab of RStudio.
}
