% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_design.R
\name{diagnose_design}
\alias{diagnose_design}
\title{Diagnose a design object}
\usage{
diagnose_design(design, population_draws = 2, sample_draws = 2,
  assignment_draws = 2, bootstrap_diagnosands = TRUE,
  population_replicates = 50)
}
\arguments{
\item{design}{Design object created by \link{declare_design}.}

\item{population_draws}{Number of simulated populations to draw.}

\item{sample_draws}{Number of samples to draw per population.}

\item{assignment_draws}{Number of random assignments to draw per sample if there is sampling or per population if there is no sampling.}
}
\value{
Diagnosis object, with simulations of estimates and estimands and estimated diagnosands.
}
\description{
Diagnose a design object
}

