% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_design.R
\name{diagnose_design}
\alias{diagnose_design}
\title{Diagnose a design object}
\usage{
diagnose_design(design, population_draws = 2, sample_draws = 2,
  assignment_draws = 2, bootstrap_diagnosands = TRUE,
  population_replicates = 1000)
}
\arguments{
\item{assignment_draws}{}
}
\description{
Diagnose a design object
}

