% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_design.R
\name{diagnose_design}
\alias{diagnose_design}
\title{Diagnose the properties of a research design}
\usage{
diagnose_design(design = NULL, diagnosis = NULL,
  statistics = list(calculate_mean_PATE, calculate_sd_PATE,
  calculate_superpopulation_RMSE, calculate_superpopulation_bias,
  calculate_superpopulation_coverage, calculate_superpopulation_type_S_rate,
  calculate_superpopulation_exaggeration_ratio, calculate_population_RMSE,
  calculate_population_bias, calculate_population_coverage,
  calculate_population_type_S_rate, calculate_population_exaggeration_ratio,
  calculate_mean_SATE, calculate_sd_SATE, calculate_sample_RMSE,
  calculate_sample_bias, calculate_sample_coverage,     
  calculate_sample_type_S_rate, calculate_sample_exaggeration_ratio,
  calculate_mean_power, calculate_sd_power, calculate_mean_estimate,
  calculate_sd_estimate), population_draws = 10, sample_draws = 5)
}
\arguments{
\item{design}{A design object created by \code{\link{declare_design}}.}

\item{diagnosis}{A diagnosis object created by \code{\link{diagnose_design}}. This allows you to calculate additional summary statistics for existing diagnosis simulations.}

\item{statistics}{A list of statistic functions that take a list of estimates, sample_estimands, and/or population_estimands and return a statistic (scalar) and a label.}

\item{population_draws}{Number of draws of the population, if using super-population.}

\item{sample_draws}{Number of draws of a sample from each population draw.}
}
\description{
Diagnose the properties of a research design
}

