% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagnose_design.R
\name{diagnose_design}
\alias{diagnose_design}
\title{Diagnose the properties of a research design}
\usage{
diagnose_design(design = NULL, diagnosis = NULL,
  statistics = list(calculate_PATE, calculate_sd_SATE, calculate_power,
  calculate_RMSE, calculate_bias, calculate_coverage, calculate_type_S_rate,
  calculate_exaggeration_ratio), sims = 5, ...)
}
\arguments{
\item{design}{}

\item{statistics}{}

\item{sims}{}

\item{...}{}

\item{simulations}{}

\item{labels}{}
}
\description{
Diagnose the properties of a research design
}

