% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/builtin_custom_functions.R
\name{difference_in_means_blocked}
\alias{difference_in_means_blocked}
\title{Built-in Estimators: Blocked Difference-in-means}
\usage{
difference_in_means_blocked(formula, data, weights = NULL, subset = NULL,
  block_variable_name = NULL, alpha = 0.05)
}
\arguments{
\item{formula}{An object of class "formula", such as Y ~ Z}

\item{data}{A data.frame, often created by \code{\link{draw_population}}.}

\item{weights}{An optional vector of weights (not yet implemented).}

\item{subset}{An optional vector specifying a subset of observations to be used.}

\item{block_variable_name}{The name of the blocking variable.}

\item{alpha}{The significance level, 0.05 by default.}
}
\description{
Built-in Estimators: Blocked Difference-in-means
}

