% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_blocks_draw_clusters.R
\name{draw_blocks}
\alias{draw_blocks}
\title{Draw blocks for custom block functions used in an assignment function}
\usage{
draw_blocks(data, assignment, random_seed = NULL)
}
\arguments{
\item{data}{}

\item{assignment}{}

\item{random_seed}{}
}
\value{

}
\description{
Draw blocks for custom block functions used in an assignment function
}

