% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_blocks_draw_clusters.R
\name{draw_clusters}
\alias{draw_clusters}
\title{Draw clusters for custom cluster functions used in an assignment function}
\usage{
draw_clusters(data, assignment)
}
\arguments{
\item{data}{A data.frame used to draw clusters from.}

\item{assignment}{An \code{assignment} object created by \code{\link{declare_assignment}}.}
}
\value{
a data.frame including a cluster indicator.
}
\description{
Draw clusters for custom cluster functions used in an assignment function
}

