% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_design.R
\name{draw_data}
\alias{draw_data}
\title{Draw the realized data from a design}
\usage{
draw_data(design)
}
\arguments{
\item{design}{A \code{design}.}
}
\value{
A data frame.
}
\description{
Draw the realized data from a design
}
\examples{
design <- declare_model(N = 25, X = rnorm(N))
df <- draw_data(design)
nrow(df)
}
