% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{draw_data}
\alias{draw_data}
\title{Draw Data from a Design}
\usage{
draw_data(design)
}
\arguments{
\item{design}{A design created by \code{\link{declare_design}}.}
}
\description{
Draw Data from a Design
}
