% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_data}
\alias{draw_data}
\title{Draw mock data based on a design}
\usage{
draw_data(design, random_seed = NULL)
}
\arguments{
\item{design}{A design object created by \code{\link{declare_design}}.}

\item{random_seed}{}
}
\description{
Draw mock data based on a design
}

