% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_data}
\alias{draw_data}
\title{Draw mock data based on a design}
\usage{
draw_data(design, random_seed = NULL)
}
\arguments{
\item{random_seed}{}

\item{population}{}

\item{sampling}{}

\item{assignment}{}

\item{potential_outcomes}{}
}
\description{
Draw mock data based on a design
}

