% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_design.R
\name{draw_estimates}
\alias{draw_estimates}
\title{Draw the realized estimates}
\usage{
draw_estimates(design)
}
\arguments{
\item{design}{A \code{design}.}
}
\value{
A tibble of estimates with estimands joined where applicable.
}
\description{
Runs the design once and returns its estimates joined to inquiries.
}
\examples{
design <- declare_model(N = 30, U = rnorm(N), Y = U) +
  declare_inquiry(mu = mean(Y)) +
  declare_estimator(Y ~ 1, .method = lm, term = "(Intercept)",
                    inquiry = "mu", label = "ols")
draw_estimates(design)
}
\seealso{
Other drawing from a design:
\code{\link[=draw_data]{draw_data()}},
\code{\link[=draw_estimands]{draw_estimands()}},
\code{\link[=get_estimates]{get_estimates()}},
\code{\link[=run_design]{run_design()}}
}
\concept{drawing from a design}
