% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_outcomes.R
\name{draw_outcome_vector}
\alias{draw_outcome_vector}
\title{Draw observed outcome (vector)}
\usage{
draw_outcome_vector(data, potential_outcomes, attrition = NULL)
}
\arguments{
\item{data}{A data.frame object}

\item{potential_outcomes}{A potential_outcomes object created by \code{\link{declare_potential_outcomes}}.}

\item{attrition}{An attrition object created by \code{\link{declare_attrition}}.}
}
\value{
a vector of the observed outcome.
}
\description{
Draw observed outcome (vector)
}

