% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_population}
\alias{draw_population}
\title{Make the full dataset or just a sample}
\usage{
draw_population(population, potential_outcomes = NULL)
}
\description{
Make the full dataset or just a sample
}

