% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_outcomes.R
\name{draw_potential_outcome_vector}
\alias{draw_potential_outcome_vector}
\title{Draw potential outcome vector}
\usage{
draw_potential_outcome_vector(data, potential_outcomes, condition_name)
}
\arguments{
\item{data}{A data.frame object}

\item{potential_outcomes}{A potential_outcomes object created by \code{\link{declare_potential_outcomes}}.}

\item{condition_name}{A vector of condition names.}
}
\value{
a vector of a potential outcome
}
\description{
Draw potential outcome vector
}

