% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_outcomes.R
\name{draw_potential_outcomes}
\alias{draw_potential_outcomes}
\title{Draw potential outcomes}
\usage{
draw_potential_outcomes(data, potential_outcomes, condition_names = NULL,
  attrition = NULL)
}
\arguments{
\item{data}{}

\item{potential_outcomes}{}

\item{condition_names}{}
}
\description{
Draw potential outcomes
}

