% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_sample_indicator}
\alias{draw_sample_indicator}
\title{draw sample from population}
\usage{
draw_sample_indicator(sampling, population_data)
}
\arguments{
\item{sampling}{A sampling object created by \code{\link{declare_sampling}}; or a function that samples}

\item{data}{A dataframe, often created by \code{\link{make_data}}.}
}
\value{
A vector of 0's and 1's indicating which population units are sampled.
}
\description{
Description
}

