% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redesign.R
\name{expand_design}
\alias{expand_design}
\title{Build a family of designs from a designer function}
\usage{
expand_design(designer, ..., expand = TRUE)
}
\arguments{
\item{designer}{A function returning a \code{design}.}

\item{...}{Named parameter values to vary.}

\item{expand}{If \code{TRUE}, expand the grid; if \code{FALSE}, zip values.}
}
\value{
A single design or a list of designs.
}
\description{
Build a family of designs from a designer function
}
\examples{
designer <- function(N) declare_model(N = N, Y = rnorm(N))
expand_design(designer, N = c(10, 20))
}
