% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redesign.R
\name{find_all_objects}
\alias{find_all_objects}
\title{Find all objects used in quosures in a design}
\usage{
find_all_objects(design)
}
\arguments{
\item{design}{A DeclareDesign design object}
}
\value{
A tibble with columns: name, value_str, step, quosure, env
}
\description{
Traverses each step in a DeclareDesign object and finds the names and environments
of all objects captured inside quosures.
}
