% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{get_balance}
\alias{get_balance}
\title{Calculate Covariate Balance}
\usage{
get_balance(data, covariates, treatment_assignment = "Z", assignment,
  report_difference = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{A dataframe, often created by \code{\link{draw_population}} or \code{\link{draw_sample}}.}

\item{covariates}{A character vector of covariate names to be used in the balance assessment.}

\item{treatment_assignment}{The name of the treatment assignment variable in data}

\item{assignment}{A assignment object, created by \code{\link{declare_assignment}}.}

\item{report_difference}{what is it?}

\item{na.rm}{what is it?}
}
\value{
A balance object
}
\description{
Description
}
\examples{
##examples go here.
}

