% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reveal_design.R
\name{get_design_probs}
\alias{get_design_probs}
\title{Calculate probabilties of assignment}
\usage{
get_design_probs(design, data)
}
\arguments{
\item{design}{A design object created by declare_design(); or a function that assigns treatment}

\item{data}{A dataframe, often created by make_data()}
}
\value{
a matrix of probabilities of assignment to treatment
}
\description{
Description
}
\examples{
# some examples go here
}

