% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_diagnosand.R
\name{get_diagnosand}
\alias{get_diagnosand}
\title{Title}
\usage{
get_diagnosand(diagnosand, simulations)
}
\arguments{
\item{diagnosand}{}

\item{na.action}{How the diagnosand summary of diagnostic statistics handles NAs. Can be \link{na.pass}, \link{na.omit}, etc.}
}
\description{
Title
}

