% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_diagnosands}
\alias{get_diagnosands}
\title{Get the diagnosands table from a diagnosis}
\usage{
get_diagnosands(diagnosis)
}
\arguments{
\item{diagnosis}{A \code{diagnosis} object.}
}
\value{
A tibble of diagnosands.
}
\description{
Get the diagnosands table from a diagnosis
}
\examples{
design <- declare_model(N = 30, Y = rnorm(N), Z = rep(0:1, 15)) +
  declare_inquiry(ATE = 0) +
  declare_estimator(Y ~ Z, .method = lm, term = "Z", inquiry = "ATE",
                    label = "ols")
d <- diagnose_design(design, sims = 5, bootstrap_sims = 0)
get_diagnosands(d)
}
