% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_design.R
\name{get_diagnosands}
\alias{get_diagnosands}
\title{Get the diagnosands data frame from a diagnosis}
\usage{
get_diagnosands(diagnosis)
}
\arguments{
\item{diagnosis}{A diagnosis object created by \code{\link{diagnose_design}}.}
}
\value{
a \code{data.frame} of the diagnosand values from a diagnosis
}
\description{
Get the diagnosands data frame from a diagnosis
}
