% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_diagnosand.R
\name{get_diagnosands}
\alias{get_diagnosands}
\title{Get Diagnosand Values}
\usage{
get_diagnosands(diagnosand, simulations, get_diagnostic_statistics = FALSE)
}
\arguments{
\item{diagnosand}{A \code{diagnosand} object created by \code{\link{declare_diagnosand}}.}

\item{simulations}{A data frame of simulations, typically created by \code{\link{diagnose_design}}.}

\item{get_diagnostic_statistics}{Flag for whether diagnostic statistics are drawn before calculating diagnosands}
}
\value{
A data.frame of diagnosands.
}
\description{
Get Diagnosand Values
}

