% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declare_estimand.R
\name{get_estimands}
\alias{get_estimands}
\title{Obtain Estimands}
\usage{
get_estimands(estimand = NULL, estimator = NULL, sample_data = NULL,
  population_data = NULL)
}
\arguments{
\item{estimand}{An estimand object or list of estimand objects, created with \code{\link{declare_estimand}}.}

\item{estimator}{An estimantor object or a list of estimator objects, created with \code{\link{declare_estimator}}.}

\item{sample_data}{A sample data.frame, created by \code{\link{draw_sample}}.}

\item{population_data}{A population data.frame, created by \code{\link{draw_population}}.}
}
\description{
Obtain Estimands
}

