% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{get_estimates}
\alias{get_estimates}
\title{Get Estimates from a Design}
\usage{
get_estimates(design)
}
\arguments{
\item{design}{A design created by \code{\link{declare_design}}.}
}
\value{
a \code{data.frame} of estimates calculated from a designs estimators.
}
\description{
Get Estimates from a Design
}
