% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_estimator.R
\name{get_estimates}
\alias{get_estimates}
\title{Get estimates}
\usage{
get_estimates(estimator, data)
}
\arguments{
\item{estimator}{An estimator object created by \code{declare_estimator}.}

\item{data}{A data.frame object, used to fit the model.}
}
\description{
Get estimates
}

