% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_probabilities.R
\name{get_sampling_probabilities}
\alias{get_sampling_probabilities}
\title{Calculate probabilties of assignment}
\usage{
get_sampling_probabilities(data, sampling)
}
\arguments{
\item{data}{A dataframe, often created by \code{\link{draw_population}} or \code{\link{draw_sample}}.}

\item{assignment}{A assignment object created by \code{\link{declare_assignment}}; or a function that assigns treatment}
}
\value{
A matrix of probabilities of assignment to treatment.
}
\description{
Description
}
\examples{
# these examples don't work yet
# smp <- declare_population(N = 850)
# po <- declare_potential_outcomes(condition_names = c("Z0","Z1"),
#                                    formula = Y ~ .01 + 0*Z0 + .2*Z1)
# assignment <- declare_assignment(potential_outcomes = po, m=200)
# mock          <- draw_population(potential_outcomes = po, sample =  smp)
# mock$Z        <- assign_treatment_indicator(assignment, data = mock)
# assignment_probs <- get_assignment_probabilities(assignment, mock)

head(assignment_probs)
}

