% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_simulations}
\alias{get_simulations}
\title{Get the simulations table from a diagnosis}
\usage{
get_simulations(diagnosis)
}
\arguments{
\item{diagnosis}{A \code{diagnosis} object.}
}
\value{
A tibble of simulations.
}
\description{
Get the simulations table from a diagnosis
}
\examples{
design <- declare_model(N = 30, Y = rnorm(N), Z = rep(0:1, 15)) +
  declare_inquiry(ATE = 0) +
  declare_estimator(Y ~ Z, .method = lm, term = "Z", inquiry = "ATE",
                    label = "ols")
d <- diagnose_design(design, sims = 5, bootstrap_sims = 0)
get_simulations(d)
}
