% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_design.R
\name{get_simulations}
\alias{get_simulations}
\title{Get the simulations data frame from a diagnosis}
\usage{
get_simulations(diagnosis)
}
\arguments{
\item{diagnosis}{A diagnosis object created by \code{\link{diagnose_design}}.}
}
\value{
a \code{data.frame} of the simulations from a diagnosis
}
\description{
Get the simulations data frame from a diagnosis
}
