% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template}
\alias{get_template}
\title{Get Design Template Function from Web}
\usage{
get_template(template_name)
}
\arguments{
\item{template_name}{Name of template to download from DeclareDesign Web site.}
}
\description{
Get Design Template Function from Web
}

