% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{insert_step}
\alias{insert_step}
\title{Insert a step into a design}
\usage{
insert_step(design, new_step, after = NULL, before = NULL)
}
\arguments{
\item{design}{A \code{design}.}

\item{new_step}{A \code{design_step} to insert.}

\item{after, before}{A label, integer, or \code{design_step} indicating the
anchor point. Provide exactly one of the two.}
}
\value{
A \code{design}.
}
\description{
Insert a step into a design
}
\examples{
design <- declare_model(N = 30, Y = rnorm(N)) +
  declare_inquiry(mu = mean(Y))
new_design <- insert_step(design,
  declare_measurement(Y2 = Y * 2),
  after = "model")
names(new_design)
}
\seealso{
Other modifying a design:
\code{\link[=delete_step]{delete_step()}},
\code{\link[=design_parameters]{design_parameters()}},
\code{\link[=expand_design]{expand_design()}},
\code{\link[=redesign]{redesign()}},
\code{\link[=replace_step]{replace_step()}}
}
\concept{modifying a design}
