% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Make the full dataset or just a sample}
\usage{
make_data(sample, potential_outcomes = NULL, assign_treatment = FALSE,
  design = NULL, outcome_variable = NULL, reveal_outcome = FALSE,
  noncompliance = NULL, sep = "_")
}
\arguments{
\item{sample}{A sample object made with \code{\link{declare_sample}}. Contains sample size, structure, custom-DGP functions, and other information related to baseline.}

\item{potential_outcomes}{An optional argument for a potential_outcomes object made with \code{\link{declare_potential_outcomes}}, or a list of potential_outcomes objects. Contains all of the necessary information to construct the potential outcomes revealed by the experiment. Potential outcomes can be a function of previous outcomes in a list of potential outcomes.}

\item{assign_treatment}{An optional argument indicating whether treatment should be assigned as part of make_data(). If TRUE, then the user must also provide arguments to design, outcome_variable, and reveal_outcome (see below).}

\item{design}{An optional argument for a design object, made with \code{\link{declare_design}}, or list of design objects. Contains the randomization procedure and other important information, such as the name of the treatment variable. Must be provided when assign_treatment == TRUE.}

\item{outcome_variable}{An optional string with the name of the outcome variable. Must be provided when assign_treatment == TRUE.}

\item{reveal_outcome}{An optional argument that reveals the observed outcome under treatment when TRUE.}

\item{noncompliance}{A noncompliance object, made with \code{\link{declare_noncompliance}}.}

\item{sep}{a character string used in the naming of potential outcomes. Defaults to "_".}
}
\description{
Make the full dataset or just a sample
}

