% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Make the full dataset or just a sample}
\usage{
make_data(potential_outcomes = NULL, sample_frame = NULL, blocks = NULL,
  clusters = NULL, N = NULL, sep = "_")
}
\arguments{
\item{potential_outcomes}{An outcomes_object made with declare_potential_outcomes()}

\item{sample_frame}{A sample_frame object made with declare_sample_frame(), or a pre-existing dataframe}

\item{N}{If sample_frame is provided, this argument is ignored.}
}
\description{
Make the full dataset or just a sample
}

