% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{model_handler}
\alias{model_handler}
\alias{tidy_estimator}
\title{Deprecated alias for label_estimator}
\usage{
model_handler(...)

tidy_estimator(...)
}
\arguments{
\item{...}{Forwarded to \code{\link[=label_estimator]{label_estimator()}}.}
}
\value{
A function.
}
\description{
\code{model_handler()} and \code{tidy_estimator()} are retained for compatibility but
will be removed in a future version. Use \code{\link[=label_estimator]{label_estimator()}} instead.
}
