% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{modify_design}
\alias{modify_design}
\title{Modify a Design}
\usage{
modify_design(design, ...)
}
\arguments{
\item{design}{A design created by \code{\link{declare_design}}.}

\item{...}{A set of modifications to a design.}
}
\description{
Modify a Design
}
