% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{modify_design}
\alias{modify_design}
\alias{add_step}
\alias{remove_step}
\alias{replace_step}
\title{Modify a Design}
\usage{
modify_design(design, ...)

add_step(..., before = NULL, after = NULL)

remove_step(...)

replace_step(..., replace)
}
\arguments{
\item{design}{a design object, usually created by \code{\link{declare_design}}, \code{\link{quick_design}}, or \code{\link{download_design}}.}

\item{...}{a series of calls to \code{\link{add_step}}, \code{\link{remove_step}}, or \code{\link{replace_step}}}

\item{before}{bare (unquoted) name of the step before which to add steps}

\item{after}{bare (unquoted) name of the step after which to add steps}

\item{replace}{bare (unquoted) name of step to be replaced}

\item{...}{bare (unquoted) names of step(s) to add to a design}

\item{...}{bare (unquoted) names of step(s) to remove from a design}

\item{...}{bare (unquoted) names of step(s) to replace in a design}
}
\value{
A design object. See \code{\link{declare_design}} for details.
}
\description{
Modify a Design
}
\section{Functions}{
\itemize{
\item \code{add_step}: 

\item \code{remove_step}: 

\item \code{replace_step}: 
}}

\examples{

 my_population <- declare_population(N = 100, noise = rnorm(N))

 my_potential_outcomes <-
   declare_potential_outcomes(Y_Z_0 = noise,
                              Y_Z_1 = noise + rnorm(N, mean = 2, sd = 2))

 my_assignment <- declare_assignment(m = 50)
 my_assignment_2 <- declare_assignment(m = 25)

 design <- declare_design(my_population,
                          my_potential_outcomes,
                          my_assignment)

 design

 modify_design(design, replace_step(my_assignment_2, replace = my_assignment))

 modify_design(design, add_step(dplyr::mutate(income = noise^2), after = my_assignment))
 modify_design(design, add_step(dplyr::mutate(income = noise^2), before = my_assignment))

 modify_design(design, remove_step(my_assignment))
}
