% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_design.R
\name{modify_design}
\alias{modify_design}
\title{Modify the design}
\usage{
modify_design(design, population = NULL, sampling = NULL,
  assignment = NULL, estimator = NULL, potential_outcomes = NULL,
  diagnosand = NULL, inputs = NULL, label = NULL, description = NULL)
}
\arguments{
\item{design}{A design object, created by \code{\link{declare_design}}.}

\item{population}{A population object, as created by \code{\link{declare_population}}.}

\item{sampling}{A sampling object, as created by \code{\link{declare_sampling}}.}

\item{assignment}{An assignment object, as created by \code{\link{declare_assignment}}.}

\item{estimator}{An estimator object, as created by \code{\link{declare_estimator}}. Optionally a list of estimator objects.}

\item{potential_outcomes}{A potential_outcomes object, as created by \code{\link{declare_potential_outcomes}}. Optionally a list of potential_outcomes objects.}

\item{diagnosand}{}

\item{inputs}{An inputs object, as created by \code{\link{declare_inputs}}.}

\item{label}{An optional design label.}

\item{description}{A description of the modified design in words.}
}
\value{
A design object
}
\description{
This function takes a design object and allows the user to modify one or more of the declarations that make up a design.
}

