% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reveal_design.R
\name{observed_outcome}
\alias{observed_outcome}
\title{Reveal observed outcomes based on a given treatment assignment}
\usage{
observed_outcome(outcome = "Y", treatment_assignment, data, sep = "_")
}
\arguments{
\item{outcome}{A character string}

\item{treatment_assignment}{A string indicating the name of the realized treatment assignment vector}

\item{data}{A data frame including the outcome and realized treatment assignment vectors indicating by outcome and treatment_assignment}

\item{sep}{The character separating outcomes from condition names in the potential outcomes columns of data}
}
\value{
an outcome vector of observed y
}
\description{
Description
}
\examples{
# some examples go here
}

