% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reveal_design.R
\name{observed_probs}
\alias{observed_probs}
\title{Reveal probabilties of assignment to realized treatment conditions}
\usage{
observed_probs(treatment_assignment, design, data)
}
\arguments{
\item{treatment_assignment}{The name of the treatment assignment variable in data.}

\item{design}{A design object created by \code{\link{declare_design}}; or a function that assigns treatment}

\item{data}{A dataframe, often created by \code{\link{make_data}}.}
}
\value{
A vector probabilities of assignment to treatment.
}
\description{
Description
}
\examples{
smp <- declare_sample(N = 850)
po <- declare_potential_outcomes(condition_names = c("Z0","Z1"),
                                   outcome_formula = Y ~ .01 + 0*Z0 + .2*Z1)
design <- declare_design(potential_outcomes = po, m=200)
mock          <- make_data(potential_outcomes = po, sample_frame =  smp)
mock$Z        <- assign_treatment(design, data = mock)
mock$prob_obs        <- observed_probs("Z", design, mock)

table(mock$prob_obs)
}

