% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcomes_and_covariates.R
\name{plot_covariates}
\alias{plot_covariates}
\title{Plot the frequency or density of covariates}
\usage{
plot_covariates(data, covariate_names, color_palette = NULL,
  barplot = FALSE, density = FALSE, histogram = FALSE, ...)
}
\arguments{
\item{data}{The dataframe that contains the outcomes to plot (i.e. made using \code{\link{draw_population}})}

\item{covariate_names}{A vector of names of the outcomes to plot}

\item{color_palette}{Optional vector of custom colors for the plot}

\item{barplot}{If TRUE, plot is a barplot}

\item{density}{If TRUE, plot is a density plot}

\item{histogram}{If TRUE, plot is a frequency histogram}

\item{...}{Optional graphical arguments}
}
\description{
Plot the frequency or density of covariates
}

