% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pop.var}
\alias{pop.var}
\title{Population variance}
\usage{
pop.var(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A scalar.
}
\description{
Computes the (uncorrected) variance: the mean squared deviation from the
mean. Useful for design-based standard errors.
}
\examples{
pop.var(c(1, 2, 3, 4, 5))
}
