% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_diagnosands.R
\name{pop.var}
\alias{pop.var}
\title{Population variance function}
\usage{
pop.var(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
numeric scalar of the population variance
}
\description{
Population variance function
}
\examples{

x <- 1:4
var(x) # divides by (n-1)
pop.var(x) # divides by n


}
