% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pre_register.R
\name{pre_register}
\alias{pre_register}
\title{Create pre-registration object to document a design, power analysis, and experiment analyses}
\usage{
pre_register(design, title, authors)
}
\arguments{
\item{design}{A design object from the declare_design function.}

\item{title}{Title of the pre-registration}

\item{authors}{List of authors, consisting of vectors of two strings -- one for the author's name and one for her affiliation. e.g. \code{list(c("Michael LaCour", "UCLA"), c("Donald Green", "Columbia University"))}. Optionally, a third item in the vector can be the text for a footnote with author contact information.}
}
\value{
\code{pre_registration} object that can be exported.
}
\description{
Create pre-registration object to document a design, power analysis, and experiment analyses
}
\examples{
export(design.object, compile_pdf = TRUE, open_pdf = TRUE)
}

