% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_diagnoses.R
\name{print.compared_diagnoses}
\alias{print.compared_diagnoses}
\title{Print a diagnosand comparison}
\usage{
\method{print}{compared_diagnoses}(x, ...)
}
\arguments{
\item{x}{A \code{compared_diagnoses} object.}

\item{...}{Passed to the data frame print method.}
}
\value{
The input invisibly.
}
\description{
Print a diagnosand comparison
}
\examples{
design <- declare_model(N = 100, Y = rnorm(N), Z = randomizr::complete_ra(N)) +
  declare_inquiry(ATE = 0) +
  declare_estimator(Y ~ Z, .method = lm, term = "Z", inquiry = "ATE",
                    label = "ols")
print(compare_diagnoses(design, redesign(design, N = 200),
                        sims = 20, bootstrap_sims = 20))
}
