% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.design}
\alias{print.design}
\title{Print a design}
\usage{
\method{print}{design}(x, ...)
}
\arguments{
\item{x}{A \code{design}.}

\item{...}{Ignored.}
}
\value{
The input invisibly.
}
\description{
Print a design
}
\examples{
d <- declare_model(N = 30, Y = rnorm(N)) + NULL
print(d)
}
