% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_dc.R
\name{print.design_comparison}
\alias{print.design_comparison}
\title{print.design_comparison
Print function for design comparison.}
\usage{
\method{print}{design_comparison}(design_comparison,
  display = c("highlights", "all", "none"))
}
\arguments{
\item{design_comparison}{A comparison generated by compare_designs.}
}
\description{
print.design_comparison
Print function for design comparison.
}
