% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_dc.R
\name{print.design_comparison}
\alias{print.design_comparison}
\title{print.design_comparison
Print function for design comparison.}
\usage{
\method{print}{design_comparison}(design_comparison,
  display = c("highlights", "all", "none"), output_type = c("html",
  "pdf", "word", "github", "Rmd", "none"),
  file_prefix = "design_comparison", render_quietly = FALSE)
}
\arguments{
\item{design_comparison}{A comparison generated by compare_designs.}

\item{display}{c("highlights", "all", "none"), where highlights is the default.}

\item{output_type}{c("html", "pdf", "word", "github", "Rmd", "none"). Creates a file called 'design_comparison.html' unless file_prefix is provided (or a different file type is specified). 'Rmd' means write R markdown only (otherwise Rmd is saved along with desired type of output file).}

\item{file_prefix}{Optional. If provided, names the file specified by 'output_type.'}

\item{render_quietly}{If FALSE (default), preview pops up and rendering details display to console.}
}
\description{
print.design_comparison
Print function for design comparison.
}
