% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.diagnosis}
\alias{print.diagnosis}
\title{Print a diagnosis}
\usage{
\method{print}{diagnosis}(x, ...)
}
\arguments{
\item{x}{A \code{diagnosis}.}

\item{...}{Ignored.}
}
\value{
The input invisibly.
}
\description{
Print a diagnosis
}
\examples{
design <- declare_model(N = 30, Y = rnorm(N), Z = rep(0:1, 15)) +
  declare_inquiry(ATE = 0) +
  declare_estimator(Y ~ Z, .method = lm, term = "Z", inquiry = "ATE",
                    label = "ols")
d <- diagnose_design(design, sims = 5, bootstrap_sims = 0)
print(d)
}
