% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_versions}
\alias{print_versions}
\title{Print version of R and packages to improve reproducibility}
\usage{
print_versions()
}
\description{
Print version of R and packages to improve reproducibility
}

