% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_design.R
\name{quick_design}
\alias{quick_design}
\title{Declare a Design via a Template Function}
\usage{
quick_design(template = NULL, expand = TRUE, ...)
}
\arguments{
\item{template}{An R function that takes a set of options and returns a design created by declare_design().}

\item{expand}{boolean - if true, form the crossproduct of the ..., otherwise recycle them}

\item{...}{Options sent to the template.}
}
\description{
Declare a Design via a Template Function
}
