% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_design_from_template.R
\name{quick_design}
\alias{quick_design}
\title{Quick Design Generation Function}
\usage{
quick_design(template, intersection = NULL, ...)
}
\arguments{
\item{template}{Template function to create a set of designs to compare}

\item{intersection}{Describes how to}

\item{...}{options for the template function. If you wish to compare designs over a range of values of a template input, use vary().}
}
\description{
Quick Design Generation Function
}

