% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_design_redesign.R
\name{redesign}
\alias{redesign}
\title{Redesign}
\usage{
redesign(design, expand = TRUE, ...)
}
\arguments{
\item{design}{a design}

\item{expand}{boolean - if true, form the crossproduct of the ..., otherwise recycle them}

\item{...}{arguments to redesign e.g., n = 100}
}
\description{
\code{redesign} quickly generates a design from an existing one by resetting symbols used in design handler parameters internally. (Advanced).
}
\details{
Importantly, \code{redesign} will edit any symbol in your design, but if the symbol you attempt to change does not exist no changes will be made and no error or warning will be issued.
}
\examples{

n <- 500
population <- declare_population(N = 1000)
sampling <- declare_sampling(n = n)
design <- population + sampling

# returns a single, modified design
modified_design <- redesign(design, n = 200)

returns a list of six modified designs
design_vary_N <- redesign(design, n = seq(400, 900, 100))

}
