% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{remove_step}
\alias{remove_step}
\title{Remove a step from an existing design}
\usage{
remove_step(...)
}
\arguments{
\item{...}{bare (unquoted) names of step(s) to remove from a design}
}
\description{
Remove a step from an existing design
}
\details{
see \code{\link{modify_design}} for details.
}
