% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{remove_step}
\alias{remove_step}
\title{Modify a Design by Removing Steps}
\usage{
remove_step(...)
}
\arguments{
\item{...}{bare (unquoted) names of steps to remove}
}
\description{
Modify a Design by Removing Steps
}
\details{
see \code{\link{modify_design}} for details.
}
