% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{replace_step}
\alias{replace_step}
\title{Replace a step in a design}
\usage{
replace_step(design, step, new_step)
}
\arguments{
\item{design}{A \code{design}.}

\item{step}{A label, integer, or \code{design_step} to replace.}

\item{new_step}{The replacement \code{design_step}.}
}
\value{
A \code{design}.
}
\description{
Replace a step in a design
}
\examples{
design <- declare_model(N = 30, Y = rnorm(N)) +
  declare_inquiry(mu = mean(Y))
replace_step(design, "mu", declare_inquiry(med = median(Y)))
}
\seealso{
Other modifying a design:
\code{\link[=delete_step]{delete_step()}},
\code{\link[=design_notes]{design_notes()}},
\code{\link[=design_parameters]{design_parameters()}},
\code{\link[=expand_design]{expand_design()}},
\code{\link[=insert_step]{insert_step()}},
\code{\link[=redesign]{redesign()}}
}
\concept{modifying a design}
