% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{replace_step}
\alias{replace_step}
\title{Modify a Design by Replacing a Step}
\usage{
replace_step(..., replace)
}
\arguments{
\item{...}{replacement step(s)}

\item{replace}{step to be replaced}
}
\description{
Modify a Design by Replacing a Step
}
\details{
see \code{\link{modify_design}} for details.
}
