% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{replace_step}
\alias{replace_step}
\title{Replace a step in an existing design}
\usage{
replace_step(..., replace)
}
\arguments{
\item{...}{bare (unquoted) names of step(s) to replace in a design}

\item{replace}{bare (unquoted) name of step to be replaced}
}
\description{
Replace a step in an existing design
}
\details{
see \code{\link{modify_design}} for details.
}
