% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_design.R
\name{replace_step}
\alias{replace_step}
\title{Replace a step in an existing design}
\usage{
replace_step(..., replace)
}
\arguments{
\item{...}{bare (unquoted) names of step(s) to replace in a design}

\item{replace}{bare (unquoted) name of step to be replaced}
}
\description{
Replace a step in an existing design
}
\details{
see \code{\link{modify_design}} for details.
}
\examples{

 my_population <- declare_population(N = 100, noise = rnorm(N))

 my_potential_outcomes <-
   declare_potential_outcomes(Y_Z_0 = noise,
                              Y_Z_1 = noise + rnorm(N, mean = 2, sd = 2))

 my_assignment <- declare_assignment(m = 50)
 my_assignment_2 <- declare_assignment(m = 25)

 design <- declare_design(my_population,
                          my_potential_outcomes,
                          my_assignment)

 design

 modify_design(design, replace_step(my_assignment_2, replace = my_assignment))

 modify_design(design, add_step(dplyr::mutate(income = noise^2), after = my_assignment))
 modify_design(design, add_step(dplyr::mutate(income = noise^2), before = my_assignment))

 modify_design(design, remove_step(my_assignment))
}
