% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reshape_diagnosis}
\alias{reshape_diagnosis}
\title{Reshape a diagnosis (returns the diagnosands table)}
\usage{
reshape_diagnosis(diagnosis, ...)
}
\arguments{
\item{diagnosis}{A \code{diagnosis} object.}

\item{...}{Reserved for future use.}
}
\value{
A tibble.
}
\description{
Reshape a diagnosis (returns the diagnosands table)
}
\examples{
design <- declare_model(N = 30, Y = rnorm(N), Z = rep(0:1, 15)) +
  declare_inquiry(ATE = 0) +
  declare_estimator(Y ~ Z, .method = lm, term = "Z", inquiry = "ATE",
                    label = "ols")
d <- diagnose_design(design, sims = 5, bootstrap_sims = 0)
reshape_diagnosis(d)
}
