% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_helper_functions.R
\name{reshape_diagnosis}
\alias{reshape_diagnosis}
\title{Clean up DeclareDesign diagnosis object for printing}
\usage{
reshape_diagnosis(diagnosis, digits = 2)
}
\arguments{
\item{diagnosis}{An object from \code{declare_design}}

\item{digits}{Number of digits.}

\item{col.names}{Allows user to provide names of columns for output. If NULL uses names from diagnosis object, if "default" uses names of default diagnosands.}

\item{n_text_fields}{Number of initial text fields in diagnosis that do not have associated standard errors.}
}
\value{
A formatted text table with bootstrapped standard errors in parentheses.
}
\description{
If diagnosands are bootstrapped, se's are put in parenthese on a second line and rounded to \code{digits}.
Function uses presence of "se(" to identify bootrapped diagnoses; avoid errors by not using "se(" in naming of diagnosands.
}
\examples{
my_population <- declare_population(N = 50, noise = rnorm(N))

my_potential_outcomes <-
  declare_potential_outcomes(Y_Z_0 = noise,
                             Y_Z_1 = noise + rnorm(N, mean = 2, sd = 2))

my_assignment <- declare_assignment(m = 25)

pate <- declare_estimand(mean(Y_Z_1 - Y_Z_0), label = "pate")

pate_estimator <- declare_estimator(Y ~ Z, estimand = pate, label = "test")


my_design <- declare_design(my_population,
                            my_potential_outcomes,
                            pate,
                            my_assignment,
                            pate_estimator)

my_diagnosis <- diagnose_design(my_design)

reshape_diagnosis(my_diagnosis)


}
