% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_helper_functions.R
\name{reshape_diagnosis}
\alias{reshape_diagnosis}
\title{Clean up DeclareDesign diagnosis object for printing}
\usage{
reshape_diagnosis(diagnosis, digits = 2, is.extracted = FALSE)
}
\arguments{
\item{diagnosis}{An object from \code{diagnose_design}, either a diagnosand dataframe or a list containing a diagnosand dataframe}

\item{digits}{Number of digits.}

\item{is.extracted}{Should be TRUE if diagnosis is a diagnosands dataframe; set to FALSE if diagnosis is a list containing the diagnosands data frame as its second element}
}
\value{
A formatted text table with bootstrapped standard errors in parentheses.
}
\description{
If diagnosands are bootstrapped, se's are put in parenthese on a second line and rounded to \code{digits}.
}
\examples{
# library(DesignLibrary)
# diagnosis <- diagnose_design(simple_two_arm_designer(), sims = 3)
# reshape_diagnosis(diagnosis)
}
