% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_diagnosands.R
\name{select_diagnosands}
\alias{select_diagnosands}
\title{Subset diagnosands by name}
\usage{
select_diagnosands(diagnosands, ...)
}
\arguments{
\item{diagnosands}{A diagnosands \code{design_step}.}

\item{...}{Names of diagnosands to keep.}
}
\value{
A diagnosands \code{design_step} containing only the selected entries.
}
\description{
Subset diagnosands by name
}
\examples{
diags <- default_diagnosands()
kept <- select_diagnosands(diags, "bias", "rmse")
names(attr(kept, "dots"))
}
