% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_diagnosands.R
\name{select_diagnosands}
\alias{select_diagnosands}
\title{Select diagnosands}
\usage{
select_diagnosands(..., alpha = 0.05, subset = NULL, na.rm = FALSE)
}
\arguments{
\item{...}{Requested diagnosands from the extended diagnosand list}

\item{alpha}{Significance level, passed to declare_diagnosands}

\item{subset}{Subset indicator, passed to declare_diagnosands}

\item{na.rm}{Logical for handling NAs when calculating diagnosands}
}
\description{
Helper to declare a set of named diagnosands
}
\details{
The diagnosands available from the extended diagnosand list are:
\preformatted{

| Diagnosand        | Meaning                                          | Formula                                                     |
|-------------------|--------------------------------------------------|-------------------------------------------------------------|
| mean_estimand     | Expected value of the estimand                   | `mean(estimand)`                                            |
| mean_estimate     | Expected value of the estimate                   | `mean(estimate)`                                            |
| bias              | (Expected) bias                                  | `mean(estimate - estimand)`                                 |
| sd_estimate       | The s.d. of the estimate ('true' standard error) | `sd(estimate)`                                              |
| rmse              | The root mean squared error                      | `sqrt(mean((estimate - estimand) ^ 2))`                     |
| power             | Power: probability of rejecting a null of zero   | `mean(p.value <= alpha)`                                    |
| coverage          | Coverage: prob the CI will cover the truth       | `mean(estimand <= conf.high & estimand >= conf.low)`        |
| mean_se           | The expected standard error                      | `mean(std.error)`                                           |
| type_s_rate       | Prob a significant estimate has wrong sign       | `mean((sign(estimate) != sign(estimand))[p.value <= alpha])`|
| exaggeration_ratio| The expected exaggeration (given significance)   | `mean((estimate/estimand)[p.value <= alpha])`               |
| var_estimate      | The variance of the estimate                     | `pop.var(estimate)`                                         |
| mean_var_hat      | Expected estimate of the 'true' var of estimate  | `mean(std.error^2)`                                         |
| prop_pos_sig      | Probability of a positive estimate given sig.    | `mean(estimate > 0 & p.value <= alpha)`                     |
| mean_ci_length    | The expected length of the confidence interval   | `mean(conf.high - conf.low)`                                |
}
}
\examples{
select_diagnosands("bias")
select_diagnosands("prop_pos_sig", "mean_se", alpha = .001)
select_diagnosands("bias")(data.frame(estimate = 2, estimand = 3))
select_diagnosands("bias")(data.frame(estimate = 1:2, estimand = c(NA, 2)))
select_diagnosands("bias", na.rm = TRUE)(data.frame(estimate = 1:2, estimand = c(NA, 2)))
}
