% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_diagnosands.R
\name{set_diagnosands}
\alias{set_diagnosands}
\title{Attach diagnosands to a design}
\usage{
set_diagnosands(design, diagnosands = default_diagnosands())
}
\arguments{
\item{design}{A \code{design}.}

\item{diagnosands}{A diagnosands \code{design_step}. Defaults to
\code{\link[=default_diagnosands]{default_diagnosands()}}.}
}
\value{
The design with \code{diagnosands} attribute set.
}
\description{
Stores \code{diagnosands} as an attribute on \code{design}; \code{\link[=diagnose_design]{diagnose_design()}} will
use these in place of the defaults when present.
}
\examples{
design <-
  declare_model(N = 30, Y = rnorm(N)) +
  declare_inquiry(mu = mean(Y)) +
  declare_estimator(Y ~ 1, .method = lm, term = "(Intercept)", inquiry = "mu")

design <- set_diagnosands(
  design,
  declare_diagnosands(mean_estimate = mean(estimate, na.rm = TRUE))
)

diagnose_design(design, sims = 5, bootstrap_sims = 0)
}
\seealso{
Other diagnosands:
\code{\link[=declare_diagnosands]{declare_diagnosands()}},
\code{\link[=default_diagnosands]{default_diagnosands()}},
\code{\link[=format.diagnosis]{format.diagnosis()}},
\code{\link[=select_diagnosands]{select_diagnosands()}}
}
\concept{diagnosands}
