% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_design.R
\name{simulate_design}
\alias{simulate_design}
\title{Run a design many times}
\usage{
simulate_design(..., sims = 500)
}
\arguments{
\item{...}{A design created by \code{\link{declare_design}}, or a set of designs. You can also provide a single list of designs, for example one created by \code{\link{expand_design}}.}

\item{sims}{The number of simulations, defaulting to 500. If sims is a vector of the form c(10, 1, 2, 1) then different steps of a design will be simulated different numbers of times.  See details.}
}
\description{
Runs many simulations of a design and saves to a dataframe.
}
\details{
Different steps of a design may each be simulated different a number of times, as specified by sims. In this case simulations are grouped into "fans", eg "fan_1" indicates all the simulations that have the same draw from the first level of the design. For efficiency there are generally fewer fans than design steps where all contiguous steps with 1 sim specified are combined into a single fan.
}
