% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_design.R
\name{simulate_single_design}
\alias{simulate_single_design}
\title{Run Single Design -- Basic simulation script that is called on by simulate_design}
\usage{
simulate_single_design(design, sims)
}
\description{
Run Single Design -- Basic simulation script that is called on by simulate_design
}
