% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{summary.design}
\alias{summary.design}
\title{Summarize a design}
\usage{
\method{summary}{design}(object, ...)
}
\arguments{
\item{object}{A \code{design}.}

\item{...}{Ignored.}
}
\value{
The input invisibly.
}
\description{
Summarize a design
}
\examples{
summary(declare_model(N = 30, Y = rnorm(N)) + NULL)
}
