% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.population}
\alias{summary.population}
\title{Summarizing population}
\usage{
\method{summary}{population}(object, extended = TRUE, stat_list = list(max =
  max, min = min, mean = mean, median = median, sd = sd), digits = 2, ...)
}
\arguments{
\item{object}{A population object as created by \code{\link{declare_population}}.}

\item{extended}{Logical. Whether to print extended summary, including variable summary statistics and levels. Default is \code{TRUE}.}

\item{stat_list}{If \code{extended = TRUE}, a list of named summary statistics, which should be reported for each variable declared in population. By default returns \code{min}, \code{max}, \code{mean}, \code{median} and \code{sd}.}

\item{digits}{Integer. Number of decimal points to prin for summary statistics. Default is 2.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
If \code{extended = FALSE} function returns list with one element, \code{summary_text}, which contains text explanation of population structure. If \code{extended = TRUE}, function returns list with 4 elements: \code{summary_text} as explained above, \code{stat}, which contains named matrix of summary statistics specified in \code{stat_list} argument, \code{code}, which contains function call which creates population object supplied, and \code{internal}, which contains list of all necessary data and options to re-create the population object using the \code{code}.
}
\description{
Summary method for population object created by \code{\link{declare_population}}.
}
\examples{
# ADD EXAMPLES HERE

}

