% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_code}
\alias{summary_code}
\title{Summary of design in code}
\usage{
summary_code(design)
}
\arguments{
\item{design}{A design object created by \code{\link{declare_design}}.}
}
\value{
character string representing the design in R code
}
\description{
Summary of design in code
}

