% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_filter.R
\name{tidy_filter}
\alias{tidy_filter}
\title{Tidy Model Results and Filter to Relevant Coefficients}
\usage{
tidy_filter(fit, term = FALSE)
}
\arguments{
\item{fit}{}

\item{term}{}
}
\value{

}
\description{
Tidy Model Results and Filter to Relevant Coefficients
}
